% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.p.directed.influences.R
\name{adjust.p.directed.influences}
\alias{adjust.p.directed.influences}
\title{Adjust p-directed influences}
\usage{
adjust.p.directed.influences(data.obj, adjust.method = c("holm", "fdr"))
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{adjust.method}{string, "holm" or "fdr"}
}
\value{
updated \code{\link{Cape}} object
}
\description{
Adjust p-directed influences
}
