% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.interactions.R
\name{allele.interactions}
\alias{allele.interactions}
\title{Makes a table of interacting alleles}
\usage{
allele.interactions(data.obj, collapsed.net = TRUE,
  color.scheme = c("CC/DO", "other"))
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{collapsed.net}{boolean, default: TRUE}

\item{color.scheme}{string "CC/DO" or "other"}
}
\value{
updated \code{\link{Cape}} object
}
\description{
Makes a table of interacting alleles
}
