% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.representation.R
\name{allele.representation}
\alias{allele.representation}
\title{Finds alleles from each founder were seleted for the pairscan}
\usage{
allele.representation(data.obj, color.scheme = c("CC/DO", "other"),
  pdf.label = "Allele.Representation.pdf")
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{color.scheme}{string "CC/DO" or "other"}

\item{pdf.label}{string}
}
\description{
Finds alleles from each founder were seleted for the pairscan
}
