% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.curve.R
\name{bin.curve}
\alias{bin.curve}
\title{Bins a singlescan curve automatically into peaks}
\usage{
bin.curve(the.curve, plot.peaks = FALSE, window.size = NULL,
  amp.min = NULL)
}
\arguments{
\item{plot.peaks}{defualt = FALSE}

\item{window.size}{}

\item{amp.min}{minimum amplitude}
}
\description{
First it finds the difference between all consecutive
points in the vector. It then looks for runs of all
positive and all negative values
it smooths the curve using a window size based on the
runs of all positive and all negative values
it then removes fluctuations in the curve that are 
smaller than the sd of the smoothed curve values
}
