% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.delta.errors.R
\name{calc.delta.errors}
\alias{calc.delta.errors}
\title{Error propogation}
\usage{
calc.delta.errors(markers, beta.m, se, beta.cov)
}
\arguments{
\item{markers}{The marker names being tested}

\item{beta.m}{The main-effects coefficient matrix for 
the pairwise regression of the given pair.}

\item{se}{The standard errors for the marker pair.}

\item{beta.cov}{The model covariance matrix from the 
pairwise regression}
}
\value{
Returns the error propagated coefficients and 
standard errors for m12 and m21
}
\description{
This function performs error propagation on 
coefficients and standard errors.
}
