% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.p.R
\name{calc.p}
\alias{calc.p}
\title{Calculate P Values for Interactions Based on Permutations}
\usage{
calc.p(data.obj, pval.correction = c("holm", "hochberg", "hommel",
  "bonferroni", "BH", "BY", "fdr", "none"))
}
\arguments{
\item{pval.correction}{One of "holm", "fdr", "lfdr" or "none", indicating 
whether the p value correction method used should be the Holm step-down procedure, 
false discovery rate, local false discovery, or no correction rate respectively.}
}
\value{
The data object is returned with a new list with two elements. The elements 
correspond to the two directions of influence: marker1 to marker2 and marker2 to 
marker1. Each element contains a table with the source and target variants, the 
empirical p values, and the adjusted p values, along with the effect size, standard 
error and t statistic for each interaction.
}
\description{
Calculate P Values for Interactions Based on Permutations
}
