% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_delta_errors.R
\name{calc_delta_errors}
\alias{calc_delta_errors}
\title{Error propagation}
\usage{
calc_delta_errors(markers, beta_m, se, beta_cov)
}
\arguments{
\item{markers}{The marker names being tested}

\item{beta_m}{The main-effects coefficient matrix for 
the pairwise regression of the given pair.}

\item{se}{The standard errors for the marker pair.}

\item{beta_cov}{The model covariance matrix from the 
pairwise regression}
}
\value{
Returns the error propagated coefficients and 
standard errors for m12 and m21
}
\description{
This function performs error propagation on 
coefficients and standard errors.
}
