% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_m.R
\name{calc_m}
\alias{calc_m}
\title{Calculate m}
\usage{
calc_m(markers, beta_m)
}
\arguments{
\item{markers}{A vector of length two holding the
names of the markers being tested.}

\item{beta_m}{The matrix of beta values from the 
pairwise linear regression testing the interaction
between the markers.}
}
\value{
This function returns a matrix with one 
row holding the m12 and m21 values as described 
in Carter et al. 2012
}
\description{
This function performs the reparameterization
step originally described in
Carter, G. W., Hays, M., Sherman, A. & Galitski, T. 
Use of pleiotropy to model genetic interactions in a 
population. PLoS Genet. 8, e1003010 (2012).
}
\keyword{internal}
