% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p.R
\name{calc_p}
\alias{calc_p}
\title{Calculate P Values for Interactions Based on Permutations}
\usage{
calc_p(
  data_obj,
  pval_correction = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none")
)
}
\arguments{
\item{data_obj}{A \code{\link{Cape}} data object}

\item{pval_correction}{One of "holm", "fdr", "lfdr" or "none", indicating 
whether the p value correction method used should be the Holm step-down procedure, 
false discovery rate, local false discovery, or no correction rate respectively.}
}
\value{
The data object is returned with a new table called var_to_var_p_val.
This table is the same as var_to_var_influences, but with p value and adjusted
p value columns appended.
}
\description{
Calculate P Values for Interactions Based on Permutations
}
