% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cape2mpp.R
\name{cape2mpp}
\alias{cape2mpp}
\title{converts a \code{\link{read.population}} object to a DO-enabled object}
\usage{
cape2mpp(data.obj, geno.obj = NULL)
}
\arguments{
\item{data.obj}{a \code{\link{read.population}} object}

\item{geno.obj}{a genotype object. If this is not supplied then it is generated here.}
}
\value{
\code{list("data.obj" = data.obj, "geno.obj" = geno.obj)}
}
\description{
This function converts an old cape object to
capeDO objects. The genotype object should ideally
be separate from the rest of the data
}
