% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_std.R
\name{center_std}
\alias{center_std}
\title{This function mean centers and standardizes a vector}
\usage{
center_std(v)
}
\arguments{
\item{v}{The vector to be mean-centered and standardized}
}
\value{
A mean-centered and standardized vector

#' @examples 
randV <- runif(10000)
hist(randV, main = "Histogram of Uniform Distribution")
centV <- center_std(randV)
hist(centV, main = "Histogram of Distribution After Mean Centering and Standardizing")
}
\description{
This function mean centers and standardizes a vector
}
