% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.underscore.R
\name{check.underscore}
\alias{check.underscore}
\title{Checks for underscores in marker names}
\usage{
check.underscore(data.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}
}
\description{
This function looks for underscores in 
marker names. Beause cape uses underscores
to separate marker from allele names, we
cannot have any in the marker names
}
