% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bad_markers.R
\name{check_bad_markers}
\alias{check_bad_markers}
\title{Checks for unused markers}
\usage{
check_bad_markers(data_obj, geno_obj)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{geno_obj}{a genotype object}
}
\description{
This function checks for markers that aren't used
in cape. For example, markers on sex chromosomes,
or mitochondrial markers. It also removes any 
markers that are invariant across all individuals.
}
\keyword{internal}
