% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_communities.R
\name{check_communities}
\alias{check_communities}
\title{Check community assignments}
\usage{
check_communities(comm)
}
\arguments{
\item{comm}{community from \code{fastgreedy.community}}
}
\description{
This function checks community assignments
and makes sure that communities are not 
split up along the chromosome
}
\keyword{internal}
