% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geno.R
\name{check_geno}
\alias{check_geno}
\title{Check to see if the any markers in the supplied
genotype matrix is one of the covariates.}
\usage{
check_geno(genotype, covar_table)
}
\arguments{
\item{genotype}{The vector of genotype values for
a single marker}

\item{covar_table}{The table of covariates}
}
\value{
covariates indices that match the genotype
}
\description{
Check to see if the any markers in the supplied
genotype matrix is one of the covariates.
}
\keyword{internal}
