% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_underscore.R
\name{check_underscore}
\alias{check_underscore}
\title{Checks for underscores in marker names}
\usage{
check_underscore(data_obj)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}
}
\description{
This function looks for underscores in 
marker names. Beause cape uses underscores
to separate marker from allele names, we
cannot have any in the marker names
}
\keyword{internal}
