% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkV.R
\name{chunkV}
\alias{chunkV}
\title{Bins a vector into chunks}
\usage{
chunkV(V, num_chunks)
}
\arguments{
\item{V}{The vector to be divided into chunks}

\item{num_chunks}{The number of chunks in which
to bin the vector.}
}
\value{
This function returns a list of length 
num_chunks. Each element in the list contains
the elements of the vector that were assigned 
to that bin.
}
\description{
This function is used to chunk a vector into 
bins. It is primarily used to divide jobs 
sent to a certain number of parallel clusters.
}
\keyword{internal}
