% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.markers.R
\name{compare.markers}
\alias{compare.markers}
\title{Removes markers from data.obj that are not present in the geno.obj}
\usage{
\method{compare}{markers}(data.obj, geno.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}
}
\value{
The data.obj is returned, and any markers that were
not present in geno.obj are removed from data.obj$geno_names
}
\description{
Removes markers from data.obj that are not present in the geno.obj
}
