% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_markers.R
\name{compare_markers}
\alias{compare_markers}
\title{Removes markers from data_obj that are not present in the geno_obj}
\usage{
compare_markers(data_obj, geno_obj)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{geno_obj}{a genotype object}
}
\value{
The data_obj is returned, and any markers that were
not present in geno_obj are removed from data_obj$geno_names
}
\description{
Removes markers from data_obj that are not present in the geno_obj
}
\keyword{internal}
