% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consec.pairs.R
\name{consec.pairs}
\alias{consec.pairs}
\title{consec.pairs}
\usage{
consec.pairs(elements)
}
\arguments{
\item{elements}{a vector of elements to be paired up}
}
\value{
a matrix with two duplicated columns of elements
}
\description{
Makes a matrix of consectutive pairs of elements
}
