% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consec_pairs.R
\name{consec_pairs}
\alias{consec_pairs}
\title{Generate a matrix of consecutive elements}
\usage{
consec_pairs(elements)
}
\arguments{
\item{elements}{a vector of elements}
}
\value{
a matrix with two columns. The first column
is identical to elements with the last element taken
off. And second column holds the next consecutive 
element.
}
\description{
Given a vector, this function generates a two-column
matrix in which each row contains an element and its
consecutive neighbor.
}
\keyword{internal}
