% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.underscore.R
\name{delete.underscore}
\alias{delete.underscore}
\title{Internal function that deletes underscores from marker names}
\usage{
\method{delete}{underscore}(data.obj, geno.obj = NULL)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}
}
\value{
lists containing "data.obj" and "geno.obj"
}
\description{
Internal function that deletes underscores from marker names
}
