% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_underscore.R
\name{delete_underscore}
\alias{delete_underscore}
\title{Delete underscores from marker names}
\usage{
delete_underscore(data_obj, geno_obj = NULL)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{geno_obj}{a genotype object}
}
\value{
lists containing \code{data_obj} and \code{geno_obj}
with underscores removed from \code{data_obj$geno_names}
and \code{dimnames(geno_obj)}. These two elements must be
entered separately into \code{\link{run_cape}}.
}
\description{
Because cape uses underscores to separate marker names
from alleles, we need to delete any underscores that 
were present in the original marker names. This function
does that.
}
\keyword{internal}
