% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.pieDO.R
\name{draw.pieDO}
\alias{draw.pieDO}
\title{Draw a pie chart at specific xy coordinates with a specified number of slices and color for each slice.}
\usage{
draw.pieDO(x = 0.5, y = 0.5, radius = 1, cols = c("red", "green"),
  border.col = "black", node.border.lwd = 1, labels = NULL,
  edges = 200, label.cex = 1, xlim = NULL, ylim = NULL,
  add = TRUE)
}
\description{
Draw a pie graph evenly divided into polygons
each colored according to cols
the center of the pie chart is at x y
and the radius is rad
}
