% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.prop.R
\name{error.prop}
\alias{error.prop}
\title{Estimate Errors of Regression Coefficients}
\usage{
error.prop(data.obj, pairscan.obj, perm = FALSE, verbose = FALSE,
  run.parallel = TRUE, n.cores = 4, just.m = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{pairscan.obj}{a pariscan object from \code{\link{pairscan}}}

\item{perm}{A logical value to indicate whether error propagation should be 
performed on the test statistics (FALSE) or the permuted test statistics (TRUE).}

\item{verbose}{A logical value to indicate whether the progress of the function 
should be printed to the screen.}

\item{run.parallel}{boolean, default = TRUE}

\item{n.cores}{integer, default = 4}

\item{just.m}{}
}
\value{
This function returns the data object with a new list element: var_to_var_influences 
if perm is set to FALSE and var.to.var.influences.perm if perm is set to TRUE. These tables 
include the errors calculated for the marker1 to marker2 influences as well as the marker2 
to marker1 influences. These results are used by \code{\link{calc.p}} to calculate empirical 
p values.
}
\description{
This function uses error propagation formulas for quantities computed from 
regression coefficients to estimate the error for all regression coefficients.
}
