% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_color_fun.R
\name{exp_color_fun}
\alias{exp_color_fun}
\title{Exponential color function}
\usage{
exp_color_fun(x_min, x_max, steepness = 1, num_cols = 256)
}
\arguments{
\item{x_min}{The minimum value assigned to a color}

\item{x_max}{The maximum value assigned to a color}

\item{steepness}{The steepness of the exponential function}

\item{num_cols}{The number of colors to generate in the ramp}
}
\value{
Numeric vector with x_min:x_max remapped onto an 
exponential curve.
}
\description{
This function maps colors onto an exponential
function for use in colors_from_values.
}
\keyword{internal}
