% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.array.R
\name{flatten.array}
\alias{flatten.array}
\title{drop the 3rd dimension of an array using a summary function, e.g., min max, mean}
\usage{
flatten.array(arrayX, margin1, margin2, slice.fun)
}
\arguments{
\item{arrayX}{3D array of values, effects or t.stats, etc.}

\item{margin1}{dimension/axis of the rows (usually 1)}

\item{margin2}{dimension/axis of the columns (usually 2)}

\item{slice.fun}{summary function, e.g., \code{function(x) mean(x, na.rm = TRUE)}}
}
\value{
an updated cape object
}
\description{
This function takes in a 3D array of values, effects or t.stats, etc.
and returns a 2D matrix in wich each entry contains a single value
representing all corresponding entries in the 3rd dimension. This
number is determined by the function the user uses. It could be min,
max, mean, etc.
}
