% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.wide.threshold.1D.R
\name{genome.wide.threshold.1D}
\alias{genome.wide.threshold.1D}
\title{Calculate a genome-wide significance threshold for the single-variant scan}
\usage{
genome.wide.threshold.1D(data.obj, geno.obj = NULL, n.perm = 100,
  scan.what = c("eigentraits", "raw.traits"), ref.allele = NULL,
  alpha = c(0.01, 0.05), model.family, run.parallel = TRUE,
  n.cores = 4)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object. If this is not supplied then it is 
generated here.}

\item{n.perm}{The number of permutations to perform. The default is 100.}

\item{scan.what}{A character value that uniquely specifies whether the 
eigentraits or phenotypes should be scanned. Options are "eigentraits"
or "raw.traits".}

\item{ref.allele}{\code{\link{Cape}} requires that one of the alleles 
in the population be selected as a reference allele. The effects of all 
alleles are then reported as the effects relative to the reference allele. 
In the DO population B represents the Black6 mouse. Because this strain is 
considered a very standard strain, the default reference allele is B. 
If another allele has a positive effect, it means that it increases the given 
phenotype relative to that in the Black6 mouse. Any other allele can be selected 
as the reference allele simply by specifying it with this argument.}

\item{alpha}{The alpha value used to calculate the threshold for which variants 
will be used as covariates in the pairwise scan.}

\item{model.family}{Indicates the model family of the phenotypes. This can be 
either "gaussian" or "binomial".}

\item{run.parallel}{boolean, default = TRUE}

\item{n.cores}{integer number of cores to use if running in parallel}
}
\description{
This function uses permtuation testing to calculate a genome-wide significance
threshold for the single-variant scan. Two user-defined thresholds are 
calculated: the lower threshold (higher alpha) can be used to determine which 
variants will be used in the pairwise scan. The higher threshold (lower alpha) 
is used to determine which variants are used as covariates in the pairwise scan. 
In each permutation, the phenotype or eigentrait is shuffled and all markers 
are retested with the permuted phenotype. The regression coefficients are 
collected from each permutation and the extreme value distribution is used to 
determine thresholds for the user-defined alpha values.
}
