% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.allele.colors.R
\name{get.allele.colors}
\alias{get.allele.colors}
\title{Makes an allele color table}
\usage{
get.allele.colors(color.scheme = c("DO/CC", "other"), alleles = NULL)
}
\arguments{
\item{color.scheme}{"DO/CC" or "other"}

\item{alleles}{an array af alleles in the set "AJ", "B6", "129", "NOD", "NZO", "CAST", "PWK", "WSB"}
}
\description{
Makes an allele color table
}
