% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.col.num.R
\name{get.col.num}
\alias{get.col.num}
\title{Find column numbers using column names}
\usage{
get.col.num(data.mat, col.which = NULL, dim.which = 2)
}
\arguments{
\item{data.mat}{a matrix with column names}

\item{col.which}{a character string indicating which column should be identified
If col.which is numeric the number is returned.}

\item{dim.which}{A number indicating whether the algorithm should look
in rows (1) or columns (2) for the name.}
}
\value{
A numeric value indicating the column number corresponding
to the name given by col.which. If col.which was a number, col.which
is returned.
}
\description{
This is an internal function to find the
column numbers of phenotypes when names
are put in instead of column numbers.
}
