% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.color.R
\name{get.color}
\alias{get.color}
\title{get a hex color string}
\usage{
get.color(col.name = c("green", "purple", "red", "orange", "blue",
  "brown", "yellow", "gray"), light.dark = c("f", "l", "d"))
}
\arguments{
\item{col.name}{string color name}

\item{light.dar}{character in ("f", "l", "d")}
}
\value{
hex color string
}
\description{
given a color name ("green", "purple", "red", "orange", "blue", "brown", "yellow", "gray")
and a darkness ("f", "l", "d"), return the coressponding hex color string.
}
