% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.concent.circ.R
\name{get.concent.circ}
\alias{get.concent.circ}
\title{This function makes a named list of concentric circles starting 
at a specified point and using a specified gap}
\usage{
get.concent.circ(list.names, start.rad = 1.05, gap.rad = 0.05)
}
\description{
This function makes a named list of concentric circles starting 
at a specified point and using a specified gap
}
