% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.covar.R
\name{get.covar}
\alias{get.covar}
\title{Get covariate information}
\usage{
get.covar(data.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}
}
\value{
Returns a list with the following elements:
covar.names: a character vector holding the names of the covariates  
covar.type: a character vector indicating whether each covariate
derived from the phenotype matrix ("p") or the genotype matrix ("g")
covar.loc: A numeric vector indicating the locations of each covaraite
covar.table: A matrix holding the individual values for each covariate.
}
\description{
This function returns information about the covariates
specified for the cape run.
}
