% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.eigentraits.R
\name{get.eigentraits}
\alias{get.eigentraits}
\title{Add eigentraits, singular values, and right singular vectors to the Cape object}
\usage{
get.eigentraits(data.obj, scale.pheno = TRUE, normalize.pheno = TRUE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{scale.pheno}{boolean, default: TRUE}

\item{normalize.pheno}{boolean, default: TRUE}
}
\value{
updated \code{\link{Cape}} object
}
\description{
This script takes in the data object and returns 
the it with eigentraits added to the data object
along with the singular values and the right
singular values.
if scale.pheno is TRUE, the phenotypes are mean
centered and standardized before the svd is run
The user also has the option to rank normalize 
the phenotypes. This argument defaults to false
because many users will do their own normalization.
}
