% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.layout.mat.R
\name{get.layout.mat}
\alias{get.layout.mat}
\title{Get the best layout matrix for a given number of panes per page.}
\usage{
get.layout.mat(num.panes, type = c("landscape", "upright"))
}
\arguments{
\item{num.panes}{The number of images to plot per page.}

\item{type}{A character vector specifying whether the layout matrix 
should have more columns than rows ("landscape"), or more rows than 
columns ("upright"), if the resulting matrix is not square.}
}
\value{
A layout matrix with positions for each element in num.panes
and zeros filling the rest of the rectangle where nothing will be plotted.
}
\description{
This function is for automatically deciding how to lay out multiple
images per page. It takes as an argument the number of images the user
wants on a single page and returns the most square matrix possible.
}
