% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.line.R
\name{get.line}
\alias{get.line}
\title{This function returns the points on a line between two points}
\usage{
get.line(x0, y0, x1, y1, dens = 5e-04)
}
\description{
This function returns the points on a line between two points
}
