% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.linearly.independent.R
\name{get.linearly.independent}
\alias{get.linearly.independent}
\title{Find linearly independent markers based on effect size}
\usage{
get.linearly.independent(data.obj, verbose = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{verbose}{boolean, default = FALSE}
}
\description{
This function is a specialized script for getting
linearly independent markers based on effect size
instead of going through a matrix in order, this
script builds a matrix of linearly independent markers
starting with markers with the highest effect size
}
