% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.marker.idx.R
\name{get.marker.idx}
\alias{get.marker.idx}
\title{This function returns the marker index for marker names or numbers}
\usage{
get.marker.idx(data.obj, markers)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers}{}
}
\description{
This function returns the marker index for marker names or numbers
}
