% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.marker.location.R
\name{get.marker.location}
\alias{get.marker.location}
\title{This function returns the chromosomal marker location for a set of markers}
\usage{
get.marker.location(data.obj, markers)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers}{}
}
\description{
This function returns the chromosomal marker location for a set of markers
}
