% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.marker.name.R
\name{get.marker.name}
\alias{get.marker.name}
\title{This function returns the marker names given the numbers}
\usage{
get.marker.name(data.obj, markers)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers}{covariate names}
}
\description{
This function returns the marker names given the numbers
}
