% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.network.R
\name{get.network}
\alias{get.network}
\title{Convert the final results to a form plotted by \code{\link{plotNetwork}} and \code{\link{plotCollapsedVarInf}}}
\usage{
get.network(data.obj, p.or.q = 0.05, min.std.effect = 0,
  standardize = FALSE, collapse.linked.markers = TRUE,
  threshold.power = 1, verbose = FALSE, plot.linkage.blocks = FALSE,
  lookup.marker.position = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{p.or.q}{A threshold indicating the maximum adjusted p value considered 
significant. If an fdr method has been used to correct for multiple testing, 
this value specifies the maximum q value considered significant.}

\item{min.std.effect}{}

\item{standardize}{A logical value. If FALSE, the interaction terms are stored 
as marker to marker influence coefficients. If TRUE, the coefficients are 
standardized by their standard errors.}

\item{collapse.linked.markers}{A logical value. If TRUE markers are combined 
into linkage blocks based on correlation. If FALSE, each marker is treated as 
an independent observation.}

\item{threshold.power}{}

\item{verbose}{default = FALSE}

\item{plot.linkage.blocks}{}

\item{lookup.marker.position}{}
}
\description{
This function collapses a network based on linkage.
It returns the data object with a new weighted 
adjacency matrix in which the weights are 
standardized effects of each influence. If there
is more than one marker in a block, the weight
is the maximum influence between markers in each
block
}
