% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pheno.R
\name{get.pheno}
\alias{get.pheno}
\title{Finds alleles from each founder were seleted for the pairscan}
\usage{
get.pheno(data.obj, covar = NULL)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{color.scheme}{string "CC/DO" or "other"}

\item{pdf.label}{string}
}
\description{
Finds alleles from each founder were seleted for the pairscan
}
