% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.stats.multiallele.R
\name{get.stats.multiallele}
\alias{get.stats.multiallele}
\title{This function performs regressions and
collects the relevant statistics.
Take out markers on the sex chromosomes}
\usage{
get.stats.multiallele(phenotype, genotype, covar.table, ph.family, ref.col)
}
\arguments{
\item{phenotype}{kinship corrected phenotype data}

\item{genotype}{kinship corrected genotype data}

\item{covar.table}{kinship corrected covariate data}

\item{ph.family}{a character string for a description of the error distribution (see? "family" param in the glm function)}

\item{ref.col}{}
}
\value{
a list with "stats", "pval", "score"
}
\description{
This function removes any markers in the geno.obj on the sex
chromosomes as well as invariant markers.
}
