% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allele_colors.R
\name{get_allele_colors}
\alias{get_allele_colors}
\title{Get DO colors}
\usage{
get_allele_colors(color_scheme = c("DO/CC", "other"), alleles = NULL)
}
\arguments{
\item{color_scheme}{"DO/CC" or "other" The default is "DO/CC"
If set to "other", colors unrelated to the DO/CC are used.}

\item{alleles}{an vector af alleles in the set "AJ", "B6", "129", 
"NOD", "NZO", "CAST", "PWK", "WSB". This argument can be used to 
retrieve colors only for a subset of alleles.}
}
\value{
A table with three columns corresponding to strain names,
strain nicknames, and the assigned color values.
}
\description{
This function returns the official DO allele colors.
as described here \url{https://compgen.unc.edu/wp/?page_id=577}
}
\examples{
allele_colors <- get_allele_colors(color_scheme = "DO/CC")
barplot(rep(1, nrow(allele_colors)), col = allele_colors[,3], 
names = allele_colors[,1], main = "DO/CC colors")
other_colors <- get_allele_colors(color_scheme = "other")
barplot(rep(1, nrow(other_colors)), col = other_colors[,3], 
names = other_colors[,1], main = "Other available colors")

}
\keyword{internal}
