% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_allele.R
\name{get_block_allele}
\alias{get_block_allele}
\title{Get allele assignments for linkage blocks}
\usage{
get_block_allele(data_obj, block_name, collapsed_net)
}
\arguments{
\item{data_obj}{The cape object}

\item{block_name}{The name of the linkage block as defined
by the name of the list called linkage_blocks_full or 
linkage_blocks_collapsed}

\item{collapsed_net}{A logical value indicating whether the 
block names to be used are from linkage_blocks_collapsed (TRUE)
or linkage_blocks_full (FALSE)}
}
\value{
A vector of the alleles contained in the linkage block
}
\description{
This is an internal function used by \code{\link{plot_full_network}}
to retrieve the allele names contained in linkage blocks for 
plotting using the appropriate colors.
}
\keyword{internal}
