% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_circle.R
\name{get_circle}
\alias{get_circle}
\title{Generate coordinates for a circle}
\usage{
get_circle(radius, center_x = 1, center_y = 1, dens = 5e-04)
}
\arguments{
\item{radius}{A numeric value giving the radius of the
circle}

\item{center_x}{The x coordinate for the center of the circle}

\item{center_y}{The y coordinate for the center of the circle}

\item{dens}{A numeric value controlling how many points are returned.
Smaller values will return more points along the circumference of the
circle}
}
\value{
This function returns a list with two elements x and y.
These are the x and y coordinates of the circle. Plotted against
each other they will plot a circle.
}
\description{
Given x,y coordinates and a radius, this function
generates points on the circumference of the circle.
This function is used by plot_network to plot cape
results in a circular layout.
}
\keyword{internal}
