% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_num.R
\name{get_col_num}
\alias{get_col_num}
\title{Find column numbers using column names}
\usage{
get_col_num(data_mat, col_which = NULL, dim_which = 2)
}
\arguments{
\item{data_mat}{a matrix with column names}

\item{col_which}{a character string indicating which column should be identified
If col_which is numeric the number is returned.}

\item{dim_which}{A number indicating whether the algorithm should look
in rows (1) or columns (2) for the name.}
}
\value{
A numeric value indicating the column number corresponding
to the name given by col_which. If col_which was a number, col_which
is returned.
}
\description{
This is an internal function to find the
column numbers of phenotypes when names
are put in instead of column numbers.
}
\keyword{internal}
