% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color.R
\name{get_color}
\alias{get_color}
\title{get a hex color string}
\usage{
get_color(
  col_name = c("green", "purple", "red", "orange", "blue", "brown", "yellow", "gray"),
  light_dark = c("f", "l", "d")
)
}
\arguments{
\item{col_name}{string color name. Must be one of "green", 
"purple", "red", "orange", "blue", "brown", "yellow", "gray"}

\item{light_dark}{character value. One of ("f", "l", "d").
"l" indicates light colors, "d" indicates dark colors, and "f"
indicates colors ranging from light to dark.}
}
\value{
a vector of length four containing the hex colors 
indicated by the parameters
}
\description{
This internal function stores colors for plotting.
Given a color name ("green", "purple", "red", "orange", 
"blue", "brown", "yellow", "gray")
and a darkness: "l" for light colors, "d" for dark colors 
and "f" for the full light/dark spectrum
this function returns the corresponding ramp of colors.
}
\keyword{internal}
