% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concent_circ.R
\name{get_concent_circ}
\alias{get_concent_circ}
\title{Generate list of concentric circles}
\usage{
get_concent_circ(list_names, start_rad = 1.05, gap_rad = 0.05)
}
\arguments{
\item{list_names}{A character vector indicating labels for each circle
There will be one circle generated for each element in list_names.}

\item{start_rad}{The radius for the innermost circle.}

\item{gap_rad}{The amount of space between each circle.}
}
\value{
A list of x,y coordinates as generated
by the function \code{\link{get_circle}} and named with 
list_names
}
\description{
This is an internal function used by \code{\link{plot_network}}
and calling get_circle to generate concentric circles starting
at a specified radius and with a specified gap between them.
These circles are used for plotting traits and chromosomes
in concentric circle plots.
}
\keyword{internal}
