% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covar.R
\name{get_covar}
\alias{get_covar}
\title{Get covariate information}
\usage{
get_covar(data_obj)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}
}
\value{
Returns a list with the following elements:
covar_names: a character vector holding the names of the covariates  
covar_type: a character vector indicating whether each covariate
derived from the phenotype matrix ("p") or the genotype matrix ("g")
covar_loc: A numeric vector indicating the locations of each covariate
covar_table: A matrix holding the individual values for each covariate.
}
\description{
This function returns information about the covariates
specified for the cape run.
}
\examples{
\dontrun{
covar_info <- get_covar(data_obj)
}

}
