% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geno.R
\name{get_geno}
\alias{get_geno}
\title{Gets the geno object}
\usage{
get_geno(data_obj, geno_obj)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{geno_obj}{a genotype object.}
}
\value{
Returns the genotype array matching 
the markers and individuals specified in 
data_obj$geno_names
}
\description{
This is an internal function returns the 
genotype matrix for scanning as defined by 
the markers and individuals specified in
}
