% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geno_dim.R
\name{get_geno_dim}
\alias{get_geno_dim}
\title{Returns which dimensions the individual, locus, and
alleles are in in the genotype object.}
\usage{
get_geno_dim()
}
\value{
Returns a vector of three named elements identifying
which dimension the individual (mouse), allele, and loci are in.
}
\description{
This is an internal function that returns the 
locations of genotype dimensions from the genotype
object. It is a relic from when DOQTL put loci in the
second dimension and alleles in the third, while R/qtl2
put loci in the third dimension and alleles in the second.
}
\keyword{internal}
