% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linearly_independent.R
\name{get_linearly_independent}
\alias{get_linearly_independent}
\title{Check selected markers for linear independence.}
\usage{
get_linearly_independent(data_obj)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}
}
\value{
This function returns a list with two elements.
independent_markers: A matrix of the markers that are 
linearly independent. 
rejected_markers: A vector indicating which markers were
removed for violating linear independence.
}
\description{
This function checks a marker matrix selected by
\code{\link{select_markers_for_pairscan}} for linear
independence by checking the correlation between
pairs of markers. If any are perfectly correlated,
only one marker from the block of perfect correlation
is kept.
}
\keyword{internal}
