% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_covar.R
\name{get_marker_covar}
\alias{get_marker_covar}
\title{Get genotype or covariate values}
\usage{
get_marker_covar(data_obj, geno_obj, marker_covar_names)
}
\arguments{
\item{data_obj}{A \code{\link{Cape}} object}

\item{geno_obj}{A genotype object}

\item{marker_covar_names}{A vector of strings 
which can contain marker names with alleles appended
or covariate names.}
}
\value{
This function returns a matrix with individuals
in rows and markers/covariates in columns.
}
\description{
In some cape results plotting functions
either marker genotypes or covariates can
be used for plotting effects on traits. 
However markers and covariates are stored in
different places so it can be annoying to 
retrieve their values.
}
\details{
This function looks both in the genotype
matrix and in covariate tables to find those
values.
}
\keyword{internal}
