% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_idx.R
\name{get_marker_idx}
\alias{get_marker_idx}
\title{Get original indices for markers}
\usage{
get_marker_idx(data_obj, markers)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{markers}{A vector of marker names}
}
\value{
A vector the same length as the input markers vector
indicating the index of each marker
}
\description{
Given a vector of marker names or numbers, this 
function returns the indices in the genotype 
matrix for each marker.
}
\keyword{internal}
