% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_location.R
\name{get_marker_location}
\alias{get_marker_location}
\title{Get marker genomic position}
\usage{
get_marker_location(data_obj, markers)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{markers}{A vector of marker names}
}
\value{
A vector the same length as the input markers vector
indicating the genomic coordinate of each marker.
}
\description{
Given a vector of marker names or numbers, this 
function returns the genomic coordinates for
each marker, not including the chromosome number,
which is retrieved using \code{\link{get_marker_chr}}.
}
