% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_name.R
\name{get_marker_name}
\alias{get_marker_name}
\title{Get marker names}
\usage{
get_marker_name(data_obj, markers)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{markers}{A vector of marker numbers}
}
\value{
A vector the same length as the input 
markers vector indicating the name of each marker
}
\description{
Given a vector of marker numbers, this function 
returns the name of each marker.
}
\examples{
\dontrun{
marker_names <- get_marker_name(data_obj, 1:10)
}

}
