% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marker_num.R
\name{get_marker_num}
\alias{get_marker_num}
\title{Get numbers for markers}
\usage{
get_marker_num(data_obj, markers)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{markers}{A vector of marker names}
}
\value{
A vector the same length as the input markers vector
indicating the number of each chromosome.
}
\description{
Given a vector of marker names or numbers, this 
function returns the numeric index of each marker.
}
\keyword{internal}
