% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.missing.geno.R
\name{impute.missing.geno}
\alias{impute.missing.geno}
\title{impute missing genotype data using k nearest neighbors}
\usage{
impute.missing.geno(data.obj, geno.obj = NULL, k = 10,
  ind.missing.thresh = 0, marker.missing.thresh = 0,
  prioritize = c("fewer", "ind", "marker"), max.region.size = NULL,
  min.region.size = NULL, run.parallel = TRUE, verbose = FALSE,
  n.cores = 2)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{k}{integer number of neighboring markers, default = 10}

\item{ind.missing.thresh}{percent of individuals that are acceptible to remove, default = 0}

\item{marker.missing.thresh}{percent of genotype markers that are acceptible to remove, default = 0}

\item{prioritize}{the basis prioritization is one of 
"fewer" = remove the fewest possible cells from the matrix
"ind" = remove the fewest possible individuals
"marker" = remove the fewest possible markers}

\item{max.region.size}{maximum number of markers to be used in calculating individual similarity}

\item{min.region.size}{minimum number of markers to be used in calculating individual similarity}

\item{run.parallel}{boolean}

\item{verbose}{boolean}

\item{n.cores}{integer number of available CPU cores to use for parallel processing}
}
\value{
an updated cape object
}
\description{
This function uses k nearest neighbors to impute missing genotype data
on a per chromosome basis prioritize lets you remove individuals before
markers, vice versa, or decide based on which will remove fewer elements.
If you have a separate data.obj and geno.obj, this script returns both 
as \code{list(data.obj, geno.obj)}. The data.obj has the updated markers and 
individuals, and the geno.obj has the imputed genetic data. These objects
must be separated after running this script. max.region.size specifies the
maximum number of markers to be used in calculating individual similarity.
There is a trade-off between the time it takes to calculate a distance
matrix for a large matrix and the time it takes to slide through the
genome imputing markers. This function does not yet support imputation of
covariates.
}
