% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiate.cape.R
\name{instantiate.cape}
\alias{instantiate.cape}
\title{This function instantiates an R6 Cape object and is meant to be called from Python
using RPy2}
\usage{
instantiate.cape(
  yaml.parameters = NULL,
  results_path = NULL,
  pheno = NULL,
  chromosome = NULL,
  marker.num = NULL,
  marker.location = NULL,
  geno.names = NULL,
  geno = NULL,
  use.kinship = FALSE
)
}
\arguments{
\item{yaml.parameters}{path to the results}

\item{pheno}{phenotype data}

\item{chromosome}{chromosomes list}

\item{marker.num}{marker data}

\item{marker.location}{marker location data}

\item{geno.names}{genotype data object}

\item{geno}{genotype data}

\item{use.kinship}{TRUE or FALSE, FALSE by default}
}
\value{
\codeCape object
}
\description{
This function instantiates an R6 Cape object and is meant to be called from Python
using RPy2
}
