% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.on.the.fly.R
\name{kinship.on.the.fly}
\alias{kinship.on.the.fly}
\title{Calculates the leave-two-out kinship matrix}
\usage{
kinship.on.the.fly(kin.obj, geno, chr1 = NULL, chr2 = NULL,
  phenoV = NULL, covarV = NULL)
}
\arguments{
\item{kin.obj}{a kinship object}

\item{geno}{a genotype object. If this is not supplied then it is generated here.}

\item{covarV}{}
}
\value{
\code{list("err.cov", "corrected.pheno", "corrected.geno", "corrected.covar")}
}
\description{
This function calculates the leave-two-out kinship matrix
used in capeRel
plot.adj.mat is used to plot the covariance matrices 
vs. the positive definite matrices
to correct just for the covariate, leave chr1 and chr2 NULL
}
