% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinship.R
\name{Kinship}
\alias{Kinship}
\title{Calculate the kinship correction matrix}
\usage{
Kinship(data.obj, geno.obj, type = c("overall", "ltco"), n.cores = 4,
  pop = c("MPP", "2PP", "RIL"))
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{type}{correction type, must be either "overall" or "ltco" (leave two chromosomes out)}

\item{n.cores}{integer, default 4}

\item{pop}{population type, "MPP" (multi-parental population), "2PP" (2 parents), "RIL" (recombinant inbred line)}
}
\description{
This function produces a realized relationship matrix (kinship matrix)
for use in adjusting for the effect of inbred relatedness 
plot.adj.mat is used to plot the covariance matrices
vs. the positive definite matrices
}
