% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.blocks.network.R
\name{linkage.blocks.network}
\alias{linkage.blocks.network}
\title{This function uses the genotype correlations, but
finds blocks within the correlation matrix}
\usage{
linkage.blocks.network(data.obj, collapse.linked.markers = TRUE,
  threshold.power = 1, plot.blocks = TRUE,
  lookup.marker.position = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{collapse.linked.markers}{A logical value. If TRUE markers are combined 
into linkage blocks based on correlation. If FALSE, each marker is treated as 
an independent observation.}

\item{threshold.power}{}

\item{plot.blocks}{default = TRUE}

\item{lookup.marker.position}{default = TRUE}

\item{p.or.q}{A threshold indicating the maximum adjusted p value considered 
significant. If an fdr method has been used to correct for multiple testing, 
this value specifies the maximum q value considered significant.}

\item{min.std.effect}{}

\item{standardize}{A logical value. If FALSE, the interaction terms are stored 
as marker to marker influence coefficients. If TRUE, the coefficients are 
standardized by their standard errors.}

\item{verbose}{default = FALSE}

\item{plot.linkage.blocks}{}
}
\description{
This function uses the genotype correlations, but
finds blocks within the correlation matrix
}
