% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.input.and.run.cape.R
\name{load.input.and.run.cape}
\alias{load.input.and.run.cape}
\title{This function loads the input file path and runs cape}
\usage{
load.input.and.run.cape(
  input_file = NULL,
  yaml_params = NULL,
  results_path = NULL,
  use_kinship = FALSE,
  run_parallel = FALSE
)
}
\arguments{
\item{input_file}{data input to be loaded}

\item{yaml_params}{a parameter set up in the form of a YAML string}

\item{results_path}{path to the results}

\item{use_kinship}{TRUE or FALSE, FALSE by default}

\item{run_parallel}{FALSE by default}
}
\description{
This function loads the input file path and runs cape
}
