% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker2covar.R
\name{marker2covar}
\alias{marker2covar}
\title{Allows users to change covariate assignments}
\usage{
marker2covar(data.obj, geno.obj = NULL, singlescan.obj = NULL,
  covar.thresh = NULL, markers = NULL)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{singlescan.obj}{a singlescan object is required if setting covariates by a t threshold.}

\item{covar.thresh}{}

\item{markers}{}
}
\description{
This script allows users to change covariate assignments
manually.
markers <- rownames(covar.flags)[5:10]
pheno <- "ET1"
It adds the information to the data object
only the output from singlescan and pairscan
make separate objects.
}
