% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_image_plot.R
\name{my_image_plot}
\alias{my_image_plot}
\title{Generate a Heatmap-type image}
\usage{
my_image_plot(x, ...)
}
\arguments{
\item{x}{Matrix to be plotted}

\item{...}{possible parameters are main, xlab, ylab, mark_coords,
mark_col, show_labels, chromosome_coordinates, chr_names, chr_labels,
show_pheno_labels, extra_col_mat,allele_cols}
}
\description{
This internal function generates heatmap-type
images for functions like \code{\link{plot_svd}}
\code{\link{plot_pairscan}}, and \code{\link{plot_variant_influences}}
}
\keyword{internal}
