% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pheno.R
\name{norm_pheno}
\alias{norm_pheno}
\title{Mean-center and normalize phenotypes}
\usage{
norm_pheno(data_obj, mean_center = TRUE)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object
mean_center a logical value indicating whether the
traits should be mean centered. If FALSE, the traits
are only normalized.}

\item{mean_center}{mean center}
}
\value{
the data object is returned. The pheno slot of
the data object will have normalized and/or mean-centered
traits. The function also preserves the original trait matrix
in a slot called raw_pheno.
}
\description{
This function is a wrapper for mean-centering
normalizing and standardizing the trait matrix.
in a data_obj.
}
\examples{
\dontrun{
data_obj <- norm_pheno(data_obj)
}

}
