% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.pairscan.parallel.R
\name{one.pairscan.parallel}
\alias{one.pairscan.parallel}
\title{Run a single pairscan}
\usage{
one.pairscan.parallel(data.obj, phenotype.vector, genotype.matrix,
  int = NULL, covar.vector = NULL, paired.markers, n.perm = 0,
  run.parallel = TRUE, verbose = FALSE, n.cores = 4)
}
\description{
This function is an internal function to run a single pairscan
It is used both to do the actual pairscan, as well as to 
do the permutations of the pairscan.
it takes in as arguments the vector of phenotype values, the
matrix of genotypes for all markers and a vector of flags for
covariates.
It returns the effects matrix and the covariance matrix 
for all marker pairs in the genotype matrix. The results
are returned in a list that can be inserted directly into
the data.obj if this is the actual scan. The results can
also be dissected for the maximum effect if the scan is
being run for a permutation test. In the case of a permutation
test, the phenotype.vector being supplied should be permuted.
}
