% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.singlescanDO.R
\name{one.singlescanDO}
\alias{one.singlescanDO}
\title{Performs a 1D scan for generating the 2D null distribution}
\usage{
one.singlescanDO(phenotype.vector, genotype.mat, model.family,
  ref.allele = "A", covar.vector = NULL, run.parallel = TRUE,
  n.cores = 4)
}
\arguments{
\item{phenotype.vector}{A vector of phenotype values, one entry for each individual.}

\item{genotype.mat}{A matrix of genotype values with individuals in rows and markers 
in columns. Matrix entries contain the probability of the reference allele at each 
position for each individual.}

\item{model.family}{Indicates the model family of the phenotypes. This can be 
either "gaussian" or "binomial".}

\item{ref.allele}{the reference allele from the Cape data.obj}

\item{covar.vector}{A vector of 1's and 0's indicating which markers should be used 
as covariats (1) and which should not (0), optional}

\item{run.parallel}{default = TRUE}

\item{n.cores}{default = 4}
}
\description{
Performs a 1D scan for generating the 2D null distribution
}
