% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.matrix.R
\name{pair.matrix}
\alias{pair.matrix}
\title{get all 2-combinations of elements in a vector}
\usage{
pair.matrix(elements, ordered = FALSE, self.pairs = FALSE)
}
\arguments{
\item{elements}{a vector containing elements from which to draw all possible pairs}

\item{ordered}{boolean, TRUE if the elements are ordered}

\item{self.pairs}{boolean, TRUE if we want to include self-pairs}
}
\description{
This script makes a matrix with two columns 
listing all pairwise combinations of the elements
of the input vector.
the input vector must have all elements you want
pairs of. It can be numeric or character strings
For example, to get all pairs of numbers between 
1 and 5, markers = c(1,2,3,4,5)
}
