% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_matrix.R
\name{pair_matrix}
\alias{pair_matrix}
\title{Get all pairs of elements in a vector}
\usage{
pair_matrix(elements, ordered = FALSE, self_pairs = FALSE)
}
\arguments{
\item{elements}{a vector containing elements from 
which to draw all possible pairs. Can be numeric or 
strings}

\item{ordered}{A logical value indicating whether the order of the
pairs is important. If TRUE both a,b and b,a will be returned.
If FALSE, only a,b will be returned.}

\item{self_pairs}{A logical value indicating whether self pairs
should be included. If TRUE, a,a will be returned. Otherwise it
will be excluded.}
}
\value{
A two-column matrix containing pairs of entries in elements.
}
\description{
This function makes a matrix with two columns 
listing all pairwise combinations of the elements
of the input vector.
}
\keyword{internal}
