% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairscan.kin.R
\name{pairscan.kin}
\alias{pairscan.kin}
\title{This function runs a pairscan when the kinship correction is being used.}
\usage{
pairscan.kin(data.obj, geno.obj = NULL, scan.what, marker.pairs, kin.obj,
  verbose = TRUE, run.parallel = TRUE, n.cores = 2)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{scan.what}{A character string uniquely identifying whether eigentraits
or raw traits should be scanned. Options are "eigentraits", "raw.traits"}

\item{marker.pairs}{all length-2 combinations of markers}

\item{kin.obj}{a kinship object}

\item{verbose}{show progress messages}

\item{run.parallel}{to parallel, or not to parallel}

\item{n.cores}{if parallel, how many cores?}
}
\description{
This function runs a pairscan when the kinship correction is being used.
}
