% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairscan.noKin.R
\name{pairscan.noKin}
\alias{pairscan.noKin}
\title{This function performs a pairscan for when there is no kinship correction required.}
\usage{
pairscan.noKin(data.obj, pheno.mat, geno.mat, covar.table, marker.pairs,
  n.perm, verbose, run.parallel = TRUE, n.cores = NULL)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{pheno.mat}{a phenotype matrix}

\item{geno.mat}{a genotype matrix}

\item{marker.pairs}{all length-2 combinations of markers}

\item{verbose}{show progress messages}

\item{run.parallel}{to parallel, or not to parallel}

\item{n.cores}{if parallel, how many cores?}
}
\description{
This function performs a pairscan for when there is no kinship correction required.
}
