% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairscan.null.kin.R
\name{pairscan.null.kin}
\alias{pairscan.null.kin}
\title{Generates a null distribution for the pairscan}
\usage{
pairscan.null.kin(data.obj, geno.obj = NULL, kin.obj = NULL,
  scan.what = c("eigentraits", "raw.traits"),
  pairscan.null.size = NULL, max.pair.cor = NULL,
  min.per.geno = NULL, model.family = "gaussian",
  marker.selection.method = c("top.effects", "uniform", "effects.dist",
  "by.gene"), run.parallel = TRUE, n.cores = 4, verbose = FALSE,
  gene.list = NULL)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{kin.obj}{a kinship object}

\item{scan.what}{A character string uniquely identifying whether eigentraits
or raw traits should be scanned. Options are "eigentraits", "raw.traits"}

\item{max.pair.cor}{A numeric value between 0 and 1 indicating the maximum
Pearson correlation that two markers are allowed. If the correlation
between a pair of markers exceeds this threshold, the pair is not tested.
If this value is set to NULL, min.per.genotype must have a numeric value.}

\item{min.per.geno}{The minimum number of individuals allowable per
genotype. If for a given marker pair, one of the genotypes is
underrepresented, the marker pair is not tested. If this value is NULL,
max.pair.cor must have a numeric value.}

\item{model.family}{Indicates the model family of the phenotypes. This can be 
either "gaussian" or "binomial".}

\item{marker.selection.method}{options are "top.effects", "uniform", "effects.dist", "by.gene"}

\item{verbose}{boolean, default = FALSE}

\item{gene.list}{boolean, only required for "by.gene" marker selection method}
}
\description{
This script generates a null distribution
for the pairscan. For each permutation,
it runs a single scan and selects the top
N markers. It then uses these markers to
perform a permutation of the pairscan.
the null distribution generated here uses
a fixed number of the TOP ranking markers
from the permuted single scan
TODO 
we need to update it to use select.markers.for.pairscan
if marker.selection.method is netwas, you need to provide
a list of genes from the netWAS analysis
}
