% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_generate_breaks.R
\name{pheatmap_generate_breaks}
\alias{pheatmap_generate_breaks}
\title{pheatmap generate breaks found at this link
https://cran.r-project.org/package=pheatmap}
\usage{
pheatmap_generate_breaks(x, n, center = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of breaks}

\item{center}{logical defaults to False}
}
\description{
This function is internal to 'pheatmap' and not exported, but 
our code depends on it. We added it here till it becomes exported.
}
\keyword{internal}
