% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheatmap_scale_colours.R
\name{pheatmap_scale_colours}
\alias{pheatmap_scale_colours}
\title{pheatmap scale colours found at this link
https://cran.r-project.org/package=pheatmap}
\usage{
pheatmap_scale_colours(mat, col = rainbow(10), breaks = NA, na_col)
}
\arguments{
\item{mat}{matrix}

\item{col}{color palette default set to rainbow(10)}

\item{breaks}{number of breaks}

\item{na_col}{color for na values}
}
\description{
This function is internal to 'pheatmap' and not exported, but 
our code depends on it. We added it here till it becomes exported.
}
\keyword{internal}
