% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink2qtl.R
\name{plink2qtl}
\alias{plink2qtl}
\title{Convert plink2 files to the QTL file format}
\usage{
plink2qtl(ped = "test.ped", map = "test.map", out = "out.csvr",
  missing.genotype = "0", no.fid = FALSE, no.parents = FALSE,
  no.sex = FALSE, no.pheno = FALSE, verbose = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{ped}{full path to the ped file}

\item{map}{full path to the map file}

\item{out}{full path to the rotated csvr file (see: http://www.rqtl.org/sampledata/). Default is out.csvr}

\item{missing.genotype}{default is "0"}

\item{no.fid}{boolean, default is FALSE}

\item{no.parents}{boolean, default is FALSE}

\item{no.sex}{boolean, default is FALSE}

\item{no.pheno}{boolean, default is FALSE}

\item{verbose}{boolean, default is FALSE, gives some happy little progress messages}

\item{overwrite}{boolean, default is FALSE, will only remove the existing file if this is set to TRUE}
}
\value{
None
}
\description{
1. Dowload the plink executable from: https://www.cog-genomics.org/plink2
2. Convert your plink file with: plink --bfile YOUR-PLINK-FILE --export ped --out PED-FILE
3. Run your PED-FILE through this script to get a file formated for R/qtl
}
