% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trait.circ.R
\name{plot.trait.circ}
\alias{plot.trait.circ}
\title{Plot concentric trait circles}
\usage{
\method{plot}{trait.circ}(trait.circ, label.gap, plot.dim, main.lwd)
}
\arguments{
\item{trait.circ}{A list with one element for each cirlce giving
the coordinates of the points to plot for each circle.}

\item{label.gap}{The amount of space to leave between the circles
and their labels}

\item{plot.dim}{the dimensions of the plot}

\item{main.lwd}{The line thickness for the circles}
}
\value{
None
}
\description{
This internal function is called by \code{\link{PlotNetwork}}
It plots one circle for each trait in the data object.
}
