% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trait_circ.R
\name{plot_trait_circ}
\alias{plot_trait_circ}
\title{Plot concentric trait circles}
\usage{
plot_trait_circ(trait_circ, label_gap, plot_dim, main_lwd)
}
\arguments{
\item{trait_circ}{A list with one element for each cirlce giving
the coordinates of the points to plot for each circle.}

\item{label_gap}{The amount of space to leave between the circles
and their labels}

\item{plot_dim}{the dimensions of the plot}

\item{main_lwd}{The line thickness for the circles}
}
\value{
None
}
\description{
This internal function is called by \code{\link{plot_network}}
It plots one circle for each trait in the data object.
}
\keyword{internal}
