% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFullNetwork.R
\name{plotFullNetwork}
\alias{plotFullNetwork}
\title{Plot the final epistatic network in a traditional network view.}
\usage{
plotFullNetwork(data.obj, p.or.q = 0.05, collapsed.net = TRUE,
  main = NULL, color.scheme = c("DO/CC", "other"), pos.col = "brown",
  neg.col = "blue", bg.col = "gray", light.dark = "l",
  node.border.lwd = 1, layout.matrix = NULL, zoom = 1, xshift = 0,
  yshift = 0, node.radius = 1, label.nodes = TRUE,
  label.offset = 0, label.cex = 1, legend.radius = 1,
  legend.cex = 1, legend.position = "topleft",
  arrow.offset = node.radius, arrow.length = 0.2, edge.lwd = 2)
}
\description{
This function plots the final results in a layout different to
 both \code{\link{plotVariantInfluences}} and \code{\link{plotNetwork}}. 
 In this view, the network is plotted with a traditional network layout. 
 The positional information in \code{\link{plotNetwork}} is lost, but 
 in this view it is easier to see the structure of the overall network 
 in terms of hub and peripheral nodes. In this view, each node is plotted 
 as a pie-chart, and the main effects of the node are indicated as 
 positive (green), negative (red), or not-significant (gray). Significant 
 interactions are shown as green (positive) or red (negative) arrows between 
 nodes. The function \code{\link{get.network}} must be run before plotting 
 the network.
}
