% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetworkDO.R
\name{plotNetworkDO}
\alias{plotNetworkDO}
\title{This script draws a head specifically over electrode locations
It shouldn't need to be scaled independently of the net layout}
\usage{
plotNetworkDO(data.obj, marker.pairs = NULL, collapsed.net = TRUE,
  trait = NULL, phenotype.labels = NULL, ref.allele = "B",
  color.scheme = c("DO/CC", "other"), main.lwd = 4, inter.lwd = 3,
  label.cex = 1.5, percent.bend = 15, chr.gap = 1, label.gap = 5,
  positive.col = "brown", negative.col = "blue", show.alleles = TRUE,
  allele.labels = NULL)
}
\description{
This script draws a head specifically over electrode locations
It shouldn't need to be scaled independently of the net layout
}
