% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairscan.R
\name{plotPairscan}
\alias{plotPairscan}
\title{Plot the result of the 2D scan}
\usage{
plotPairscan(data.obj, pairscan.obj, phenotype = NULL,
  standardized = FALSE, show.marker.labels = FALSE, show.chr = TRUE,
  label.chr = TRUE, show.alleles = TRUE, allele.labels = NULL,
  pos.col = "brown", neg.col = "blue", color.scheme = c("DO/CC",
  "other"), pdf.label = "Pairscan.Regression.pdf")
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{pairscan.obj}{a pariscan object from \code{\link{pairscan}}}

\item{phenotype}{default = NULL, can be drawn from \code{pairscan.obj$pairscan.results}}

\item{standardized}{default = FALSE}

\item{pos.col}{default = "brown"}

\item{neg.col}{default = "blue"}

\item{color.scheme}{options are "DO/CC" and "other"}

\item{pdf.label}{default = "Pairscan.Regression.pdf"}
}
\description{
This script plots the results of the 2D scan.
It plots a 2D matrix of the t-statistics of
the interactions between all pairs of 
markers.
The data object is only used to get marker
order. This might be redundant, but it seemed
easier for the user to put this in then to 
extract the marker order themselves
}
