% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSinglescan.R
\name{plotSinglescan}
\alias{plotSinglescan}
\title{Makes effects plots for multi-allelic 1D scans.}
\usage{
plotSinglescan(data.obj, singlescan.obj, chr = NULL, traits = NULL,
  alpha = c(0.01, 0.05), standardized = TRUE,
  color.scheme = c("DO/CC", "other"), allele.labels = NULL,
  include.covars = TRUE, show.selected = FALSE, line.type = "l",
  lwd = 1, pch = 16, cex = 1, covar.label.size = 0.7)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{singlescan.obj}{a singlescan object}

\item{chr}{an array of chromosome names to include in the plot}

\item{traits}{an array of traits to include in the plot}

\item{alpha}{the alpha significance level}

\item{standardized}{boolean; singlescan.t.stats if TRUE, else singlescan.effects}

\item{color.scheme}{"DO/CC" or "other"}

\item{allele.labels}{an array of character labels}

\item{include.covars}{boolean}

\item{show.selected}{boolean; if TRUE, run select.markers.for.pairscan() first}

\item{line.type}{1-character string giving the type of plot desired. The following values are possible, for details, see plot: "p" for points, "l" for lines, "b" for both points and lines, "c" for empty points joined by lines, "o" for overplotted points and lines, "s" and "S" for stair steps and "h" for histogram-like vertical lines. Finally, "n" does not produce any points or lines.}

\item{lwd}{line width, default is 1}

\item{pch}{see the "points()" R function. Default is 16 (a point).}

\item{cex}{see the "points()" R function. Default is 1.}

\item{covar.label.size}{default is 0.7}
}
\description{
Makes effects plots for multi-allelic 1D scans.
}
