% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSVD.R
\name{plotSVD}
\alias{plotSVD}
\title{Plots the restuls of the singular value decomposition on the phenotypes}
\usage{
plotSVD(data.obj, orientation = c("vertical", "horizontal"),
  neg.col = "blue", pos.col = "brown", light.dark = "f",
  pheno.labels = NULL, cex.barplot.axis = 1.7,
  cex.barplot.labels = 2, cex.barplot.title = 1.7,
  main = "Eigentrait Contributions to Phenotypes", cex.main = 2,
  main.x = 0.5, main.y = 0.5, cex.ET = 1.7, ET.label.x = 0.5,
  ET.label.y = 0.5, pheno.label.pos = 0.5, cex.pheno = 1.7,
  pheno.srt = 90, percent.total.variance.x = 0.5,
  percent.total.variance.y = 0.5, cex.color.scale = 1,
  cex.var.accounted = 2, var.accounted.x = 0, var.accounted.y = 0,
  show.var.accounted = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{orientation}{string, ("vertical", "horizontal")}

\item{neg.col}{(default: "blue")}

\item{pos.col}{(default: "brown")}

\item{light.dark}{(default: "f")}

\item{pheno.labels}{array of strings}

\item{cex.barplot.axis}{(default: 1.7)}

\item{cex.barplot.labels}{(default: 2)}

\item{cex.barplot.title}{(default: 1.7)}

\item{main}{(default: "Eigentrait Contributions to Phenotypes")}

\item{cex.main}{(default: 2)}

\item{main.x}{(default: 0.5)}

\item{main.y}{(default: 0.5)}

\item{cex.ET}{(default: 1.7)}

\item{ET.label.x}{(default: 0.5)}

\item{ET.label.y}{(default: 0.5)}

\item{pheno.label.pos}{(default: 0.5)}

\item{cex.pheno}{(default: 1.7)}

\item{pheno.srt}{(default: 90)}

\item{percent.total.variance.x}{(default: 0.5)}

\item{percent.total.variance.y}{(default: 0.5)}

\item{cex.color.scale}{(default: 1)}

\item{cex.var.accounted}{(default: 2)}

\item{var.accounted.x}{(default: 0)}

\item{var.accounted.y}{(default: 0)}

\item{show.var.accounted}{boolean}
}
\value{
\code{list("data.obj" = data.obj, "geno.obj" = geno.obj)}
}
\description{
This script plots the results of the singular value
decomposition on the phenotypes. When a user has more
than two phenotypes, they may wish to use fewer eigentraits
than are returned by default. 
It takes in the data object after the svd has been done
It creates a raster plot of the contribution of each 
eigentrait to each phenotype with a barplot on the 
side of how much variance each eigentrait captures
The script also plots the log10 of the singular values.
The user can then use the script select.eigentraits to
modify the data object to only carry the eigentraits of
interest.
orientation determines whether the plot is laid out 
vertically or horizontally. The default is vertical.
}
