% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariantInfluences.R
\name{plotVariantInfluences}
\alias{plotVariantInfluences}
\title{This function plots m12 and m21 using myImagePlot}
\usage{
plotVariantInfluences(data.obj, p.or.q = 0.05, min.std.effect = 0,
  plot.all.vals = FALSE, all.markers = FALSE, standardize = TRUE,
  color.scheme = c("DO/CC", "other"), pos.col = "brown",
  neg.col = "blue", not.tested.col = "lightgray",
  show.marker.labels = FALSE, show.chr = TRUE, label.chr = TRUE,
  show.alleles = TRUE, scale.effects = c("log10", "sqrt", "none"),
  pheno.width = 11, covar.width = 11, covar.labels = NULL,
  phenotype.labels = NULL, show.not.tested = TRUE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{p.or.q}{A threshold indicating the maximum adjusted p value considered 
significant. If an fdr method has been used to correct for multiple testing, 
this value specifies the maximum q value considered significant.}

\item{min.std.effect}{}

\item{plot.all.vals}{}

\item{pos.col}{}

\item{neg.col}{}

\item{not.tested.col}{}

\item{show.marker.labels}{}

\item{show.chr}{}

\item{label.chr}{}

\item{show.alleles}{}

\item{scale.effects}{}

\item{pheno.width}{width of the phenotype text on the plot, in pixels}

\item{covar.width}{width of the covariate text on the plot, in pixels}

\item{covar.labels}{}

\item{phenotype.labels}{}

\item{show.not.tested}{}
}
\description{
This function plots the reparameterized influences of variants on each other. 
The epistatic interactions from the pairwise scan are reparameteriezed to 
the terms \eqn{m_{12}} and \eqn{m_{21}}, where the subscripts indicate the 
source and target variants respectively. These terms are interpreted as the 
effect that the source variant exerts on the target variant when both are 
present. Negative influences represent suppression while positive influences 
represent enhancement.
}
