% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnorm_pheno.R
\name{qnorm_pheno}
\alias{qnorm_pheno}
\title{Plot trait distributions}
\usage{
qnorm_pheno(data_obj, pheno_which = NULL, pheno_labels = NULL)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{pheno_which}{A vector of trait names to plot. The default is to plot all traits.}

\item{pheno_labels}{A vector of names for traits to appear in the plot in case the column names are not very pretty.}
}
\description{
This function plots the quantiles of each trait
against quantiles of a theoretical normal distribution.
This provides a way to check whether traits are normally
distributed
}
