% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.parameters.R
\name{read.parameters}
\alias{read.parameters}
\title{Read the parameter file, add missing entries}
\usage{
read.parameters(filename = "cape.parameters.yml")
}
\arguments{
\item{filename}{full path to the .yml file holding CAPE parameters}
}
\value{
Returns a named list with all possible options
}
\description{
This function returns reads in the YAML file
and checks for any parameters that might not
be included. This may not matter for the given
run, but it's handy to be able to check for 
any and all potential variables.
}
