% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parameters.R
\name{read_parameters}
\alias{read_parameters}
\title{Read the parameter file, add missing entries}
\usage{
read_parameters(filename = "cape.parameters.yml", yaml_parameters = NULL)
}
\arguments{
\item{filename}{full path to the .yml file holding CAPE parameters (is not needed if yaml_parameters is provided)}

\item{yaml_parameters}{yaml string holding CAPE parameters (can be NULL)}
}
\value{
Returns a named list with all possible options
}
\description{
This function returns reads in the YAML file
and checks for any parameters that might not
be included. This may not matter for the given
run, but it's handy to be able to check for 
any and all potential variables.
}
\examples{
\dontrun{
param_table <- read_parameters()
}

}
