% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.ind.R
\name{remove.ind}
\alias{remove.ind}
\title{This function removes individuals}
\usage{
remove.ind(data.obj, ind.to.remove = NULL, names.to.remove = NULL)
}
\arguments{
\item{ind.to.remove}{}

\item{names.to.remove}{}

\item{cape.obj}{a \code{\link{Cape}} object}
}
\value{
an updated cape data object
}
\description{
This function removes individuals
}
