% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.kin.ind.R
\name{remove.kin.ind}
\alias{remove.kin.ind}
\title{Removes individuals from the kinship object to match the cape.obj}
\usage{
remove.kin.ind(data.obj, kin.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{kin.obj}{a kinship object}
}
\value{
updated kinship object
}
\description{
Removes individuals from the kinship object to match the cape.obj
}
