% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.missing.genotype.data.R
\name{remove.missing.genotype.data}
\alias{remove.missing.genotype.data}
\title{Removes missing marker (column) or individual (row) data}
\usage{
remove.missing.genotype.data(data.obj, geno.obj = NULL,
  ind.missing.thresh = 0, marker.missing.thresh = 0,
  prioritize = c("fewer", "ind", "marker"))
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{ind.missing.thresh}{percent of individuals that are acceptible to remove, default = 0}

\item{marker.missing.thresh}{percent of genotype markers that are acceptible to remove, default = 0}

\item{prioritize}{the basis prioritization is one of 
"fewer" = remove the fewest possible cells from the matrix
"ind" = remove the fewest possible individuals
"marker" = remove the fewest possible markers}
}
\value{
an updated cape object
}
\description{
This function removes missing data and let you remove 
individuals before markers, vice versa, or 
decide based on which will remove fewer elements.
because get.geno uses geno.names in data.obj to
retrieve the genotype.object, nothing needs to be
removed from the geno.obj
}
