% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.unused.markers.R
\name{remove.unused.markers}
\alias{remove.unused.markers}
\title{Take out markers on the sex chromosomes}
\usage{
remove.unused.markers(data.obj, geno.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}
}
\value{
an updated \code{\link{Cape}} object (data.obj)
}
\description{
This function removes any markers in the geno.obj on the sex
chromosomes as well as invariant markers.
}
