% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_kin_ind.R
\name{remove_kin_ind}
\alias{remove_kin_ind}
\title{Removes individuals from the kinship object to match the cape.obj}
\usage{
remove_kin_ind(data_obj, kin_obj)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{kin_obj}{a kinship object}
}
\value{
updated kinship object
}
\description{
Removes individuals from the kinship object to match the cape.obj
}
\examples{
\dontrun{
kin_obj <- remove_kin_id(data_obj, kin_obj)
}

}
