% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_markers.R
\name{remove_markers}
\alias{remove_markers}
\title{Removes genetic markers}
\usage{
remove_markers(data_obj, markers_to_remove)
}
\arguments{
\item{data_obj}{a \code{\link{Cape}} object}

\item{markers_to_remove}{A vector of marker names to be removed.}
}
\description{
Removes genetic markers
}
