% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_progress.R
\name{report_progress}
\alias{report_progress}
\title{Report Progress of a Process}
\usage{
report_progress(
  current,
  total,
  percent_text = 10,
  percent_dot = 2,
  verbose = TRUE
)
}
\arguments{
\item{current}{The current index being processed}

\item{total}{The total number of indices that will be processed}

\item{percent_text}{The percent progress at which the percent should
be printed to the screen.}

\item{percent_dot}{The percent progress at which a dot should be printed
to the screen}

\item{verbose}{A logical value indicating whether to report progress to 
the screen. Defaults to TRUE.}
}
\value{
None. Prints output to the screen.
}
\description{
This function prints out the percent progress in a process
given the current iteration the total number of iterations
and the percentage at which a progress report is wanted.
It only works on loops in which the index is explicitly incremented
in the code.
}
\keyword{internal}
