% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.cape.R
\name{run.cape}
\alias{run.cape}
\title{Runs the CAPE algorithm}
\usage{
run.cape(data.obj, geno.obj, p.or.q = 0.05, path = ".",
  results.file = "cross.RData", n.cores = 4, run.singlescan = TRUE,
  run.pairscan = TRUE, error.prop.coef = TRUE,
  error.prop.perm = TRUE, initialize.only = FALSE, verbose = TRUE,
  run.parallel = TRUE)
}
\arguments{
\item{data.obj}{the S4 class from \code{\link{Cape}}}

\item{geno.obj}{the genotype object
TODO results.dir or resutls.file????}

\item{verbose}{boolean, output goes to stdout}

\item{run.parallel}{boolean, if TRUE runs certain parts of the code as parallel blocks}

\item{results.dir}{a full path string to an existing empty directory. An error is thrown if the directory is not empty.}
}
\value{
None, output artifacts are saved to the results.dir directory
}
\description{
This function assumes you already have all required libraries and functions loaded.
}
