% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rz_transform.R
\name{rz_transform}
\alias{rz_transform}
\title{Rank Z normalize}
\usage{
rz_transform(x, jitter = FALSE)
}
\arguments{
\item{x}{data}

\item{jitter}{boolean, default is FALSE.}
}
\value{
Returns a numeric vector with a normal distribution
}
\description{
Rank Z normalize data in a vector. We rank the data in 
x, divide by n-1, and return the quantiles of the normal
distribution using qnorm. This transformation is also 
called an inverse normal transformation.
}
\keyword{internal}
