% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.region.R
\name{segment.region}
\alias{segment.region}
\title{Divide a region into equal parts.}
\usage{
segment.region(region.min, region.max, num.points,
  alignment = c("center", "ends"))
}
\arguments{
\item{region.min}{A numerical value indicating the minimum value of the region.}

\item{region.max}{A numerical value indicating the maximum value of the region.}

\item{num.points}{The number of points to place in the region.}

\item{alignmnet}{A character element indicating whether the points should be 
centered within the region or whether they should extend to the ends of the region.}
}
\description{
This script returns n evenly spaced points in a given
region. The points can be aligned to the ends of 
the interval, or centered in the interval.
}
