% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_region.R
\name{segment_region}
\alias{segment_region}
\title{Divide a region into equal parts.}
\usage{
segment_region(
  region_min,
  region_max,
  num_points,
  alignment = c("center", "ends")
)
}
\arguments{
\item{region_min}{A numerical value indicating the minimum value of the region.}

\item{region_max}{A numerical value indicating the maximum value of the region.}

\item{num_points}{The number of points to place in the region.}

\item{alignment}{A character element indicating whether the points should be 
centered within the region or whether they should extend to the ends of the region.}
}
\value{
Returns n points spaced evenly across the defined region
}
\description{
This is an internal function used to segment regions 
for plotting. It returns n evenly spaced points in a given
region. The points can be aligned to the ends of 
the interval, or centered in the interval.
}
\keyword{internal}
