% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.markers.for.pariscan.by.gene.R
\name{select.markers.for.pairscan.by.gene}
\alias{select.markers.for.pairscan.by.gene}
\title{Select SNPs for pairscan based on a gene list}
\usage{
\method{select}{markers.for.pairscan.by.gene}(data.obj, geno.obj,
  gene.list, bp.buffer = 1000, organism = c("mouse", "human"))
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{gene.list}{an array of Ensembl gene names}

\item{bp.buffer}{default = 1000}

\item{organism}{"mouse" or "human"}
}
\description{
This function selects SNPs to test for the CAPE
pairscan based on an ordered gene list
It 	goes down list of genes from top to bottom 
pulling out SNPs from around the named genes 
until we accumulate the target number of SNPs
This function also adds the distribution of 
effect sizes to the data.obj, so you can select
by this distribution for the pairscan null
}
