% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.markers.for.pairscan.dist.R
\name{select.markers.for.pairscan.dist}
\alias{select.markers.for.pairscan.dist}
\title{Selects markers for pairscan based on effect sizes}
\usage{
\method{select}{markers.for.pairscan.dist}(data.obj, singlescan.obj,
  geno.obj, verbose = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{singlescan.obj}{a single scan object}

\item{geno.obj}{a genotype object}

\item{verbose}{boolean, default = FALSE}
}
\description{
This function takes in a singlescan object
and selects markers for the pairscan based on
the effect sizes of the markers originally selected
for the pairscan
Currently, this function is only used in generating a null
distribution, not in selecting markers initially. So it is
essentially not used anymore. At some point, I'd like to use
is for initial marker selection as well.
TODO !!!This function does not yet handle multi-parent crosses!!!
}
