% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.markers.for.pairscan.R
\name{select.markers.for.pairscan}
\alias{select.markers.for.pairscan}
\title{This function takes in a singlescan object and bins the markers into bins based on the  either the LOD score, or the curve formed by the maximum allele t score at each point}
\usage{
\method{select}{markers.for.pairscan}(data.obj, singlescan.obj, geno.obj,
  specific.markers = NULL, num.alleles = 50, peak.density = 0.5,
  window.size = NULL, tolerance = 5, plot.peaks = FALSE,
  verbose = FALSE, pdf.filename = "Peak.Plots.pdf")
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{singlescan.obj}{a singlescan object}

\item{geno.obj}{a genotype object}

\item{peak.density}{default = 0.5}

\item{pdf.filename}{default = "Peak.Plots.pdf"}
}
\value{
updated \code{\link{Cape}} object
}
\description{
This function takes in a singlescan object
and bins the markers into bins based on the 
either the LOD score, or the curve formed by 
the maximum allele t score at each point
num.alleles is the total number of alleles desired
peak.density is the density at which alleles will
be sampled within a peak. A value of 0.2 means about
20% of the alleles will be selected, while a value of 
0.5 means that 50% of alleles will be selected.
they are sampled uniformly at random
This function sets a cutoff such that you get the 
if you are specifying markers, you do not need to 
include a singlescan object.
Makes effects plots for multi-allelic 1D scans.
}
