% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.markers.for.pairscan.uniform.R
\name{select.markers.for.pairscan.uniform}
\alias{select.markers.for.pairscan.uniform}
\title{Select markers for pairscan randomly with uniform distribution}
\usage{
\method{select}{markers.for.pairscan.uniform}(data.obj, geno.obj,
  required.markers = NULL, num.alleles = 500, verbose = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}

\item{required.markers}{optional, a list of markers that must be included}

\item{num.alleles}{default = 500}

\item{verbose}{boolean, default = FALSE}
}
\description{
This function selects markers for the pairscan
uniformly at random. I'm hoping that this minimizes
the skew in the m12/m21 null distribution. 
It has the option of forcing
specific markers to be included for replication of
previous results, etc.
}
