% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.pheno.R
\name{select.pheno}
\alias{select.pheno}
\title{This function selects the phenotypes in a Cape object}
\usage{
\method{select}{pheno}(data.obj, pheno.which, min.entries = 5,
  scale.pheno = FALSE, rank.norm.pheno = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{pheno.which}{array of covariate (phenotype) names from the parameters YAML file}

\item{min.entries}{minimum number of data entries the phenotype needs to have for it to be valid}

\item{scale.pheno}{if TRUE then phenotypes are mean centered and standardized}

\item{rank.norm.pheno}{if TRUE then phenotypes are rank Z normalized}
}
\value{
updated \code{\link{Cape}} object
}
\description{
Updates the pheno object to include only `pheno.which` columns, and optionally
scales and/or normalizes them. This function is only applicable to numeric phenotypes.
}
