% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlescan.R
\name{singlescan}
\alias{singlescan}
\title{Runs a 1D scan of the phenotype data to assess which markers to use as covariates for each phenotype}
\usage{
singlescan(data.obj, geno.obj, kin.obj = NULL, n.perm = 100,
  alpha = c(0.01, 0.05), model.family = "gaussian",
  run.parallel = TRUE, n.cores = 4, verbose = FALSE,
  overwrite.alert = TRUE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object. If this is not supplied then it is generated here.}

\item{kin.obj}{a kinship object}

\item{n.perm}{integer number of permutations (default=100)}

\item{alpha}{real significance level (either 0.01 or 0.05)}

\item{model.family}{array indicating the model family of the phenotypes. This can be either "gaussian" or "binomial" if length 1, all phenotypes will be assigned to the same family. If the phenotypes are a mixture, model.family can be a vector of length Np, where Np is the number of phenotypes indicating which phenotype belongs in which family.}

\item{run.parallel}{boolean}

\item{n.cores}{integer number of cores to use if running in parallel}

\item{verbose}{boolean}

\item{overwrite.alert}{boolean, for use when the script is run from the command line}
}
\value{
\code{list("data.obj" = data.obj, "geno.obj" = geno.obj)}
}
\description{
TODO Can Anna update this description?
This script performs a 1D scan of the data
and can accept names to use as covariates.
covariates must be encoded as markers (see create.covar.R)
This script first calls the permutation script
genome.wide.significance.R to calculate the significance
threshold. It then uses this threshold to determine which
markers will be used as covariates for each phenotype
}
\details{
The script creates a table of effect sizes for each marker
for each phenotype, and a table of covariate flags indicating
which markers will be used as covariates for each phenotoype
These are both added to the data object

The user has the choice to scan the eigentraits (default)
or the original phenotypes.

ref.allele is used in a multi-allele scan. The ref.allele is
removed from the regression, and all reported effects are in
referece to the ref.allele

Threshold.fun and threshold.param work together to threshold
the "significant" loci. threshold.fun might be "t.stat.thresh.sd",
for example, which would threshold the t statistics by standard
deviation. The threshold.param that would go with this, is how
many standard deviations away from the mean you'd like to set
as the cutoff.
model.family indicates the model family of the phenotypes
This can be either "gaussian" or "binomial"
if length 1, all phenotypes will be assigned to the same
family. If the phenotypes are a mixture, model.family can
be a vector of length Np, where Np is the number of phenotypes
indicating which phenotype belongs in which family.
}
