% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortByThenBy.R
\name{sortByThenBy}
\alias{sortByThenBy}
\title{Sort a table by a list of columns}
\usage{
sortByThenBy(tableX, sort.cols = c(1, 2), col.type = c("c", "n"),
  decreasing = FALSE, return.order = FALSE)
}
\arguments{
\item{tableX}{a table object}

\item{sort.cols}{integer, number of columns to sort on (can be either 1 or 2)}

\item{col.type}{options are "c" (character) or "n" (number)}

\item{decreasing}{default = FALSE}

\item{retrun.order}{defualt = FALSE}
}
\description{
This script takes in a table and sorts by columns with each
successive column in the sort.cols vector sorted within the
factors of the previous column. Only the final column can 
contain continuous variables. All others should be discrete.
col.type defines whether each column contains a number ("n")
or a character ("c")
if return order is TRUE, the function returns the order of
each column instead of the ordered table
}
