% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePopulation.R
\name{writePopulation}
\alias{writePopulation}
\title{Save the cross data in a CSV file}
\usage{
writePopulation(data.obj, geno.obj, ref.allele = "A", na = NA,
  filename = "capeData.csv")
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object. If this is not supplied then it is generated here.}

\item{ref.allele}{a character, e.g., "A", that represents the reference allele in the data object}

\item{na}{either NA or a character used to represent a missing data value in the output}

\item{filename}{absolute or relative path to the output file's destination}
}
\description{
This function writes out a cape cross
in a csv format so you can modify a
cross and write it out in a nice format
Gets the geno object whether it is put in separately or with the data.obj
}
