% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVariantInfluences.R
\name{writeVariantInfluences}
\alias{writeVariantInfluences}
\title{This function writes out the significant influences as an edge list}
\usage{
writeVariantInfluences(data.obj, p.or.q = 0.05,
  include.main.effects = TRUE, filename = "Variant.Influences.csv",
  delim = ",", mark.covar = FALSE, write.file = TRUE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{p.or.q}{A threshold indicating the maximum adjusted p value considered 
significant. If an fdr method has been used to correct for multiple testing, 
this value specifies the maximum q value considered significant.}

\item{include.main.effects}{default = TRUE}

\item{filename}{A character vector specifying the name of the file.}

\item{delim}{A character string indicating the delimeter in the data file. 
The default indicates a comma-separated file (",").}

\item{mark.covar}{A logical value. If TRUE, an asterisk is appended the 
names of markers used as covariates in the pair scan.}

\item{write.file}{A logical value indicating whether the table should be written to a file or simply returned.}
}
\description{
This function takes in the final data object and writes the variant 
influences that are at or below the specified significance level to a 
file in the current working directory.
}
