% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.emp.p.R
\name{calc.emp.p}
\alias{calc.emp.p}
\title{Calculate empirical p-values}
\usage{
calc.emp.p(obs.dist, null.dist)
}
\description{
This function uses ecdf to calculate empirical p values
given a null distribution and an observed distribution
}
