% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cape.R
\docType{class}
\name{Cape-class}
\alias{Cape-class}
\alias{getPheno,Cape-method}
\alias{setPheno<-,Cape-method}
\alias{setPheno,Cape-method}
\alias{getChromosome,Cape-method}
\alias{setChromosome<-,Cape-method}
\alias{setChromosome,Cape-method}
\alias{getMarkerNum,Cape-method}
\alias{setMarkerNum<-,Cape-method}
\alias{setMarkerNum,Cape-method}
\alias{getMarkerLocation,Cape-method}
\alias{setMarkerLocation<-,Cape-method}
\alias{setMarkerLocation,Cape-method}
\alias{getGenoNames,Cape-method}
\alias{setGenoNames<-,Cape-method}
\alias{setGenoNames,Cape-method}
\alias{getRefAllele,Cape-method}
\alias{setRefAllele<-,Cape-method}
\alias{set,Cape-method}
\alias{getGeno,Cape-method}
\alias{setGeno<-,Cape-method}
\alias{setGeno,Cape-method}
\title{The CAPE data object}
\usage{
getPheno(x)

\S4method{getPheno}{Cape}(x)

setPheno(x) <- value

\S4method{setPheno}{Cape}(x) <- value

getChromosome(x)

\S4method{getChromosome}{Cape}(x)

setChromosome(x) <- value

\S4method{setChromosome}{Cape}(x) <- value

getMarkerNum(x)

\S4method{getMarkerNum}{Cape}(x)

setMarkerNum(x) <- value

\S4method{setMarkerNum}{Cape}(x) <- value

getMarkerLocation(x)

\S4method{getMarkerLocation}{Cape}(x)

setMarkerLocation(x) <- value

\S4method{setMarkerLocation}{Cape}(x) <- value

getGenoNames(x)

\S4method{getGenoNames}{Cape}(x)

setGenoNames(x) <- value

\S4method{setGenoNames}{Cape}(x) <- value

getRefAllele(x)

\S4method{getRefAllele}{Cape}(x)

setRefAllele(x) <- value

\S4method{setRefAllele}{Cape}(x) <- value

getGeno(x)

\S4method{getGeno}{Cape}(x)

setGeno(x) <- value

\S4method{setGeno}{Cape}(x) <- value
}
\arguments{
\item{value}{a phenotype matrix}

\item{value}{a chromosome character list}

\item{value}{An integer list of marker numbers along a chromosome}

\item{value}{A numeric list of positions in centiMorgans}

\item{value}{A list of character names for each genotype, e.g., c("A", "B")}

\item{value}{A character from the geno_names that represents the wild type}

\item{value}{An array where the dimension names must be "sample", "allele", and "locus"}
}
\description{
Class \code{Cape} defines a CAPE analysis object.
}
\section{Slots}{

\describe{
\item{\code{pheno}}{A phenotype matrix}

\item{\code{chromosome}}{A chromosome character list}

\item{\code{marker_num}}{An integer list of marker numbers along a chromosome}

\item{\code{marker_location}}{A numeric list of positions in centiMorgans}

\item{\code{geno_names}}{A list of character names for each genotype, e.g., c("A", "B")}

\item{\code{ref_allele}}{A character from the geno_names that represents the wild type}

\item{\code{geno}}{An array where the dimension names must be "sample", "allele", and "locus"}

\item{\code{parameters}}{TODO mayhap we should change this?}
}}

