% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cape.R
\docType{data}
\name{Cape-class}
\alias{Cape-class}
\alias{Cape}
\title{The CAPE data object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cape
}
\description{
Class \code{Cape} defines a CAPE analysis object.
}
\section{Slots}{

\describe{
\item{\code{pheno}}{A phenotype matrix}

\item{\code{chromosome}}{A chromosome character list}

\item{\code{marker_num}}{An integer list of marker numbers along a chromosome}

\item{\code{marker_location}}{A numeric list of positions in centiMorgans}

\item{\code{geno_names}}{A list of character names for each genotype, e.g., c("A", "B")}

\item{\code{geno}}{An array where the dimension names must be "sample", "allele", and "locus"}

\item{\code{ref_allele}}{A character from the geno_names that represents the wild type}

\item{\code{parameters}}{TODO mayhap we should change this?}

\item{\code{covar_table}}{a matrix of}

\item{\code{flat_geno}}{a flattened genotype matrix}

\item{\code{non_allelic_covar}}{covariate}
}}

\keyword{datasets}
