% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cape.R
\name{Cape-class}
\alias{Cape-class}
\alias{Cape}
\title{The CAPE data object}
\description{
Class \code{Cape} defines a CAPE analysis object.
}
\section{Slots}{

\describe{
\item{\code{parameter_file}}{string, full path to YAML file with initialization
parameters}

\item{\code{yaml_parameters}}{string representing YAML CAPE parameters}

\item{\code{results_path}}{string, full path to directory for storing results
(optional, a directory will be created if one is not specified)}

\item{\code{save_results}}{boolean, default: TRUE}

\item{\code{use_saved_results}}{boolean, default: TRUE}

\item{\code{orginism}}{options are "mouse" or "human"}

\item{\code{pheno}}{A phenotype matrix}

\item{\code{chromosome}}{A chromosome character list}

\item{\code{marker_num}}{An integer list of marker numbers along a chromosome}

\item{\code{marker_location}}{A numeric list of positions in centiMorgans}

\item{\code{marker_selection_method}}{options are "top.effects", "from.list",
"uniform", "by.gene", "effects.dist"}

\item{\code{bp_buffer}}{when the marker selection method is "by.gene" this finds genes 
from Ensemble that are within this number of base pairs of the input marker's 
position}

\item{\code{geno_names}}{A list of character names for each genotype, e.g., c("A","B")}

\item{\code{geno}}{An array where the dimension names must be "sample", "allele", and "locus"}

\item{\code{effect_size_cutoff}}{from select.markers.for.pairscan.R}

\item{\code{peak_density}}{from select.markers.for.pariscan.R}

\item{\code{window_size}}{from select.markers.for.pariscan.R}

\item{\code{tolerance}}{from select.markers.for.pairscan.R}

\item{\code{ref_allele}}{A character from the geno_names that represents the wildtype}

\item{\code{covar_table}}{a matrix of}

\item{\code{flat_geno}}{a flattened genotype matrix}

\item{\code{non_allelic_covar}}{covariate}

\item{\code{max_pair_cor}}{the maximum Pearson correlation that two markers are allowed}

\item{\code{min_per_genotype}}{minimum number of individuals allowable per genotype}

\item{\code{scan_what}}{string, "eigentraits", "normalized.traits" or "raw.traits"}

\item{\code{ET}}{eigentraits}

\item{\code{traits_scaled}}{boolean}

\item{\code{traits_normalized}}{boolean}

\item{\code{var_to_var_influences_perm}}{added in \code{\link{error.prop}} 
if error propogation is performed on permuted test statistics}

\item{\code{var_to_var_influences}}{added in \code{\link{error.prop}} 
if error propogation is performed on un-permuted test statistics}

\item{\code{pval_correction}}{options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none"}

\item{\code{linkage_blocks_collapsed}}{see \code{\link{get.network}}}

\item{\code{linkage_blocks_full}}{see \code{\link{get.network}}}

\item{\code{var_to_var_p_val}}{see \code{\link{get.network}}}

\item{\code{max_var_to_pheno_influence}}{see \code{\link{get.network}}}

\item{\code{collapsed_net}}{see \code{\link{get.network}}}

\item{\code{full_net}}{see \code{\link{get.network}}}

\item{\code{use_kinship}}{boolean}
}}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-assign_parameters}{\code{Cape$assign_parameters()}}
\item \href{#method-check_inputs}{\code{Cape$check_inputs()}}
\item \href{#method-check_geno_names}{\code{Cape$check_geno_names()}}
\item \href{#method-new}{\code{Cape$new()}}
\item \href{#method-plot_svd}{\code{Cape$plot_svd()}}
\item \href{#method-plot_singlescan}{\code{Cape$plot_singlescan()}}
\item \href{#method-plot_pairscan}{\code{Cape$plot_pairscan()}}
\item \href{#method-plot_variant_influences}{\code{Cape$plot_variant_influences()}}
\item \href{#method-plot_network_do}{\code{Cape$plot_network_do()}}
\item \href{#method-plot_full_network}{\code{Cape$plot_full_network()}}
\item \href{#method-write_variant_influences}{\code{Cape$write_variant_influences()}}
\item \href{#method-set_pheno}{\code{Cape$set_pheno()}}
\item \href{#method-set_geno}{\code{Cape$set_geno()}}
\item \href{#method-create_covar_table}{\code{Cape$create_covar_table()}}
\item \href{#method-save_rds}{\code{Cape$save_rds()}}
\item \href{#method-read_rds}{\code{Cape$read_rds()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign_parameters"></a>}}
\subsection{Method \code{assign_parameters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$assign_parameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_inputs"></a>}}
\subsection{Method \code{check_inputs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$check_inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_geno_names"></a>}}
\subsection{Method \code{check_geno_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$check_geno_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$new(
  parameter_file = NULL,
  yaml_parameters = NULL,
  results_path = NULL,
  save_results = TRUE,
  use_saved_results = TRUE,
  organism = NULL,
  pheno = NULL,
  chromosome = NULL,
  marker_num = NULL,
  marker_location = NULL,
  bp_buffer = NULL,
  geno_names = NULL,
  geno = NULL,
  .geno_for_pairscan = NULL,
  effect_size_cutoff = NULL,
  peak_density = NULL,
  window_size = NULL,
  tolerance = NULL,
  ref_allele = NULL,
  covar_table = NULL,
  flat_geno = NULL,
  non_allelic_covar = NULL,
  num_alleles_in_pairscan = NULL,
  max_pair_cor = NULL,
  min_per_genotype = NULL,
  pairscan_null_size = NULL,
  p_covar = NULL,
  g_covar = NULL,
  p_covar_table = NULL,
  g_covar_table = NULL,
  model_family = NULL,
  scan_what = NULL,
  ET = NULL,
  singular_values = NULL,
  right_singular_vectors = NULL,
  traits_scaled = NULL,
  traits_normalized = NULL,
  var_to_var_influences_perm = NULL,
  var_to_var_influences = NULL,
  pval_correction = NULL,
  var_to_var_p_val = NULL,
  max_var_to_pheno_influence = NULL,
  full_net = NULL,
  use_kinship = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_svd"></a>}}
\subsection{Method \code{plot_svd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$plot_svd(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_singlescan"></a>}}
\subsection{Method \code{plot_singlescan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$plot_singlescan(
  filename,
  singlescan.obj,
  width = 20,
  height = 6,
  units = "in",
  res = 300,
  standardized = TRUE,
  allele.labels = NULL,
  alpha = c(0.05, 0.01),
  include.covars = TRUE,
  line.type = "l",
  pch = 16,
  cex = 0.5,
  lwd = 3,
  traits = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_pairscan"></a>}}
\subsection{Method \code{plot_pairscan()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$plot_pairscan(
  filename,
  pairscan.obj,
  phenotype = NULL,
  show.marker.labels = TRUE,
  show.alleles = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_variant_influences"></a>}}
\subsection{Method \code{plot_variant_influences()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$plot_variant_influences(
  filename,
  width = 10,
  height = 7,
  p.or.q = p.or.q,
  standardize = FALSE,
  not.tested.col = "lightgray",
  covar.width = 30,
  pheno.width = 30
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_network_do"></a>}}
\subsection{Method \code{plot_network_do()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$plot_network_do(
  filename,
  label.gap = 10,
  label.cex = 1.5,
  show.alleles = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_full_network"></a>}}
\subsection{Method \code{plot_full_network()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$plot_full_network(
  filename,
  zoom = 1.2,
  node.radius = 0.3,
  label.nodes = TRUE,
  label.offset = 0.4,
  label.cex = 0.5,
  bg.col = "lightgray",
  arrow.length = 0.1,
  layout.matrix = "layout_with_kk",
  legend.position = "topright",
  edge.lwd = 1,
  legend.radius = 2,
  legend.cex = 0.7,
  xshift = -1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_variant_influences"></a>}}
\subsection{Method \code{write_variant_influences()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$write_variant_influences(filename, p.or.q = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_pheno"></a>}}
\subsection{Method \code{set_pheno()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$set_pheno(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_geno"></a>}}
\subsection{Method \code{set_geno()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$set_geno(val)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_covar_table"></a>}}
\subsection{Method \code{create_covar_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$create_covar_table(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_rds"></a>}}
\subsection{Method \code{save_rds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$save_rds(object, filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_rds"></a>}}
\subsection{Method \code{read_rds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cape$read_rds(filename)}\if{html}{\out{</div>}}
}

}
}
