% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cape2mpp.R
\name{cape2mpp}
\alias{cape2mpp}
\title{Converts a \code{\link{read.population}} object to a multi-parent object}
\usage{
cape2mpp(data.obj, geno.obj = NULL)
}
\arguments{
\item{data.obj}{a data.obj formatted for cape 1.0}

\item{geno.obj}{a genotype object. If geno.obj is NULL
the genotype object is generated from data.obj$geno.}
}
\value{
This function returns a list with two objects:
\code{list("data.obj" = data.obj, "geno.obj" = geno.obj)}
These two objects must be separated again to run through 
cape.
}
\description{
This function converts an object formatted for cape 1.0
to an object formatted for cape 2.0
}
