% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cape2mpp.R
\name{cape2mpp}
\alias{cape2mpp}
\title{Converts a \code{\link{read_population}} object to a multi-parent object}
\usage{
cape2mpp(data_obj, geno_obj = NULL)
}
\arguments{
\item{data_obj}{a data_obj formatted for cape 1.0}

\item{geno_obj}{a genotype object. If geno_obj is NULL
the genotype object is generated from data_obj$geno.}
}
\value{
This function returns a list with two objects:
\code{list("data_obj" = data_obj, "geno_obj" = geno_obj)}
These two objects must be separated again to run through 
cape.
}
\description{
This function converts an object formatted for cape 1.0
to an object formatted for cape 2.0
}
