% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_std.R
\name{center_std}
\alias{center_std}
\title{This function mean centers and standardizes a vector}
\usage{
center_std(v)
}
\arguments{
\item{v}{The vector to be mean-centered and standardized}
}
\value{
A mean-centered and standardized vector
}
\description{
This function mean centers and standardizes a vector
}
\keyword{internal}
