% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.bad.markers.R
\name{check.bad.markers}
\alias{check.bad.markers}
\title{Checks for unused markers}
\usage{
check.bad.markers(data.obj, geno.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{}
}
\description{
This function checks for markers that aren't used
in cape. For example, markers on sex chromosomes,
or mitochonrdial markers. It also removes any 
markers that are invariant across all individuals.
}
