% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.geno.R
\name{check.geno}
\alias{check.geno}
\title{Check to see if the any markers in the supplied
genotype matrix is one of the covariates.}
\usage{
check.geno(genotype, covar.table)
}
\arguments{
\item{genotype}{The vector of genotype values for
a single marker}

\item{covar.table}{The table of covariates}
}
\value{
covariates indices that match the genotype
}
\description{
Check to see if the any markers in the supplied
genotype matrix is one of the covariates.
}
