% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.markers.R
\name{compare.markers}
\alias{compare.markers}
\title{Removes markers from cape.obj that are not present in the geno.obj}
\usage{
compare.markers(cape.obj, geno.obj)
}
\arguments{
\item{cape.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object. If this is not supplied then it is generated here.}
}
\value{
\code{list("cape.obj" = cape.obj, "geno.obj" = geno.obj)}
}
\description{
Removes markers from cape.obj that are not present in the geno.obj
}
