% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.markers.R
\name{compare.markers}
\alias{compare.markers}
\title{Removes markers from data.obj that are not present in the geno.obj}
\usage{
\method{compare}{markers}(data.obj, geno.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object. If this is not supplied then it is generated here.}
}
\value{
\code{list("data.obj" = data.obj, "geno.obj" = geno.obj)}
}
\description{
Removes markers from data.obj that are not present in the geno.obj
}
