% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.circle.R
\name{get.circle}
\alias{get.circle}
\title{This function returns points on a circle given the radius, and center coordinates}
\usage{
get.circle(radius, center.x = 1, center.y = 1, dens = 5e-04)
}
\description{
This function returns points on a circle given the radius, and center coordinates
}
