% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.col.num.R
\name{get.col.num}
\alias{get.col.num}
\title{Find the column number of a pheno type from it's name}
\usage{
get.col.num(data.mat, col.which = NULL, dim.which = 2)
}
\arguments{
\item{data.mat}{a labeled matrix of phenotype with columns as phenotypes}

\item{col.which}{a character string name of a phenotype column header}

\item{dim.which}{in the absence of a column header, use an integer here to select a column}
}
\value{
None, output artifacts are saved to the results.dir directory
}
\description{
This is an internal function to find the
column numbers of phenotypes when names
are put in instead of column numbers
}
