% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.concent.circ.R
\name{get.concent.circ}
\alias{get.concent.circ}
\title{Generate list of concentric circles}
\usage{
get.concent.circ(list.names, start.rad = 1.05, gap.rad = 0.05)
}
\arguments{
\item{list.names}{A character vector indicating labels for each circle
There will be one circle generated for each element in list.names.}

\item{start.rad}{The radius for the innermost circle.}

\item{The}{amount of space between each circle.}
}
\value{
A list of x,y coordinates as generated
by the function \link{\code{get.circle}} and named with 
list.names
}
\description{
This is an internal function used by \link{\code{PlotNetwork}}
and calling get.circle to generate concentric circles starting
at a specified radius and with a specified gap between them.
These circles are used for plotting traits and chromosomes
in concentric circle plots.
}
