% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.covar.R
\name{get.covar}
\alias{get.covar}
\title{Get the phenotype-genotype covariate table}
\usage{
get.covar(data.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}
}
\value{
Returns a list with "covar.names", "covar.type", "covar.loc", "covar.table"
}
\description{
This function returns a covariate table that combines
the covariates specified from phenotypes and those
specified from genotypes
}
