% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.geno.R
\name{get.geno}
\alias{get.geno}
\title{Gets the geno object}
\usage{
get.geno(data.obj, geno.obj)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object.}
}
\value{
Returns the genotype array matching 
the markers and individuals specified in 
data.obj$geno_names
}
\description{
This is an internal function returns the 
genotype matrix for scanning as defined by 
the markers and individuals specified in
}
