% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.marker.chr.R
\name{get.marker.chr}
\alias{get.marker.chr}
\title{This function returns the marker chromosome number with covariates 
returning Chr 0}
\usage{
get.marker.chr(data.obj, markers, character.names = TRUE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers}{covariate names}
}
\description{
This function returns the marker chromosome number with covariates 
returning Chr 0
}
