% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.marker.idx.R
\name{get.marker.idx}
\alias{get.marker.idx}
\title{Get original indices for markers}
\usage{
get.marker.idx(data.obj, markers)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers}{A vector of marker names}
}
\value{
A vector the same length as the input markers vector
indicating the index of each marker
}
\description{
Given a vector of marker names or numbers, this 
function returns the indices in the genotype 
matrix for each marker.
}
