% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.marker.location.R
\name{get.marker.location}
\alias{get.marker.location}
\title{Get marker genomic position}
\usage{
get.marker.location(data.obj, markers)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers}{A vector of marker names}
}
\value{
A vector the same length as the input markers vector
indicating the genomic coordinate of each marker.
}
\description{
Given a vector of marker names or numers, this 
function returns the genomic coordinates for
each marker, not including the chromosome number,
which is retrieved using \code{\link{get.marker.chr}}.
}
