% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.marker.num.R
\name{get.marker.num}
\alias{get.marker.num}
\title{This function takes in a vector of marker names and returns their number}
\usage{
get.marker.num(data.obj, markers)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers}{covariate names}
}
\description{
This function takes in a vector of marker names and returns their number
}
