% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allele_colors.R
\name{get_allele_colors}
\alias{get_allele_colors}
\title{Get DO colors}
\usage{
get_allele_colors(color_scheme = c("DO/CC", "other"), alleles = NULL)
}
\arguments{
\item{color_scheme}{"DO/CC" or "other" The default is "DO/CC"
If set to "other", colors unrelated to the DO/CC are used.}

\item{alleles}{an vector af alleles in the set "AJ", "B6", "129", 
"NOD", "NZO", "CAST", "PWK", "WSB". This argument can be used to 
retrieve colors only for a subset of alleles.}
}
\value{
A table with three columns corresponding to strain names,
strain nicknames, and the assigned color values.
}
\description{
This function returns the official DO allele colors.
as described here:
\url{http://www.csbio.unc.edu/CCstatus/index.py?run=AvailableLines.information}
}
\keyword{internal}
