% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno2covar.R
\name{pheno2covar}
\alias{pheno2covar}
\title{Create a marker variable from a covariate phenotype}
\usage{
pheno2covar(data.obj, pheno.which)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{pheno.which}{array of covariate (phenotype) names from the parameters YAML file}
}
\description{
This script takes a variable from the phenotype matrix
for example, diet treament or sex and creates a marker
variable that can be used as a covariate.
It creates a marker that is numeric and assigns the 
numeric value to each of the alleles at all loci for 
the given individual.
}
