% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trait.circ.R
\name{plot.trait.circ}
\alias{plot.trait.circ}
\title{This function plots the trait circles which have gaps at
the right-hand sice with labels}
\usage{
plot.trait.circ(trait.circ, label.gap, plot.dim, main.lwd)
}
\description{
This function plots the trait circles which have gaps at
the right-hand sice with labels
}
