% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.markers.R
\name{remove.markers}
\alias{remove.markers}
\title{Removes genetic markers}
\usage{
remove.markers(data.obj, markers.to.remove)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{markers.to.remove}{A vector of marker names to be removed.}
}
\description{
Removes genetic markers
}
