% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.unused.markers.R
\name{remove.unused.markers}
\alias{remove.unused.markers}
\title{Take out markers not used in cape}
\usage{
remove.unused.markers(data.obj, geno.obj, verbose = FALSE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object}
}
\value{
an updated \code{\link{Cape}} object (data.obj)
}
\description{
This function removes any markers that are not
used in cape. This includes markers on the sex 
chromosomes, mitochondrial markers, and any 
invariant markers.
}
