% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.cape.R
\name{run.cape}
\alias{run.cape}
\title{Runs the CAPE algorithm}
\usage{
run.cape(data.obj, geno.obj, results.file = "cross.RData",
  p.or.q = 0.05, snp.file = NULL, n.cores = 4,
  run.singlescan = TRUE, run.pairscan = TRUE, error.prop.coef = TRUE,
  error.prop.perm = TRUE, initialize.only = FALSE, verbose = TRUE,
  run.parallel = TRUE)
}
\arguments{
\item{data.obj}{the S4 class from \code{\link{Cape}}}

\item{geno.obj}{the genotype object}

\item{results.file}{the name of the saved data.obj RData file. The base name is used as the base name for all saved RData files.}

\item{p.or.q}{A threshold indicating the maximum adjusted p value considered}

\item{snp.file}{a one column, new-line separated list of marker names that has a non-empty interectsion with the genotype names}

\item{n.cores}{integer, default is 4}

\item{run.singlescan}{boolean, defaul: TRUE}

\item{run.pairscan}{boolean, default: TRUE}

\item{error.prop.coef, }{boolean, default: TRUE}

\item{error.prop.perm, }{boolean, default: TRUE}

\item{initialize.only, }{boolean, default: FALSE}

\item{verbose}{boolean, output goes to stdout}

\item{run.parallel}{boolean, if TRUE runs certain parts of the code as parallel blocks}
}
\value{
None, output artifacts are saved to the data.obj$results_path directory
}
\description{
This function assumes you already have all required libraries and functions loaded.
}
