% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.eigentraits.R
\name{select.eigentraits}
\alias{select.eigentraits}
\title{Assign selected eigentraits in the Cape object}
\usage{
\method{select}{eigentraits}(data.obj, traits.which = c(1, 2))
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{traits.which}{integer or array of integers selecting the eigentrait column}
}
\value{
updated \code{\link{Cape}} object
}
\description{
This script is used to select individual eigentraits
after viewing the results of the svd
The script defaults to eigentraits 1 and 2
}
