% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlescan.R
\name{singlescan}
\alias{singlescan}
\title{Runs a 1D scan of the phenotype data to assess which markers to use as covariates for each phenotype}
\usage{
singlescan(data.obj, geno.obj, kin.obj = NULL, n.perm = 100,
  alpha = c(0.01, 0.05), model.family = "gaussian",
  run.parallel = TRUE, n.cores = 4, verbose = FALSE,
  overwrite.alert = TRUE)
}
\arguments{
\item{data.obj}{a \code{\link{Cape}} object}

\item{geno.obj}{a genotype object. If this is not supplied then it is generated here.}

\item{kin.obj}{a kinship object}

\item{n.perm}{integer number of permutations (default=100)}

\item{alpha}{real significance level (either 0.01 or 0.05)}

\item{model.family}{array indicating the model family of the phenotypes. This can be either "gaussian" or "binomial" if length 1, all phenotypes will be assigned to the same family. If the phenotypes are a mixture, model.family can be a vector of length Np, where Np is the number of phenotypes indicating which phenotype belongs in which family.}

\item{run.parallel}{boolean, default = TRUE}

\item{n.cores}{integer number of cores to use if running in parallel}

\item{verbose}{boolean}

\item{overwrite.alert}{boolean, for use when the script is run from the command line}
}
\value{
\code{list("data.obj" = data.obj, "geno.obj" = geno.obj)}
}
\description{
This script performs marker regression to associate
individual markers with either traits or eigentraits.
If n.perm is greater than 0, permutations are run to 
determine effect size thresholds for the alpha values
provided. The default alpha values are 0.05 and 0.01.
Covariates are specified in the cape parameter file.
}
\details{
model.family indicates the model family of the phenotypes
This can be either "gaussian" or "binomial"
if length 1, all phenotypes will be assigned to the same
family. If the phenotypes are a mixture, model.family can
be a vector of length Np, where Np is the number of phenotypes
indicating which phenotype belongs in which family of distributions.
}
